Simulate Responses from a Multivariate Linear Model

Simulate one or more sets of responses from a Multivariate Linear Model (mlm) object.

Usage

# S3 method for mlm
simulate(object, nsim = 1, seed = NULL, ...)

Arguments

object

a mlm object, typically the result of calling lm where the response is a matrix.

nsim

number of replicate datasets to simulate. If nsim is greater than 1, the output is arranged in a 3D array.

seed

an object specifying if and how the random number generator should be initialized (‘seeded’). Either NULL or an integer that will be used in a call to set.seed before simulating the response vectors. If set, the value is saved as the "seed" attribute of the returned value. The default, NULL will not change the random generator state, and return .Random.seed as the "seed" attribute, see ‘Value’

...

additional optional arguments.

Value

A matrix of simulated values for the response (or an array, for nsim greater than 1)

Details

A simulate function for mlm objects, which simulates one or more sets of responses from a Multivariate Linear Model (mlm) object. If multiple sets of responses are requested, they are returned in a 3D array, with simulation number along the third dimension.

The weights argument is currently ignored -- a constant variance-covariance matrix assumed for mlm.

See also

Author

David Warton <david.warton@unsw.edu.au>

Examples

# fit a mlm to iris data:
data(iris)
iris.mlm=lm(cbind(Sepal.Length,Sepal.Width,Petal.Length,Petal.Width)~Species,data=iris)
# simulate new responses:
simulate(iris.mlm)
#>     Sepal.Length Sepal.Width Petal.Length  Petal.Width
#> 1       4.744855    3.191742    0.9314286 -0.290441877
#> 2       4.437416    2.930117    0.9273577  0.237318617
#> 3       5.059899    3.588233    1.5766764  0.241565773
#> 4       4.749934    2.711674    1.3683270 -0.012806432
#> 5       4.699386    3.590567    1.2751894  0.273295736
#> 6       4.915199    3.807838    1.1308123  0.330121196
#> 7       5.844763    3.535132    2.0167076  0.115148357
#> 8       5.148641    3.412047    1.7017840  0.433898007
#> 9       5.176379    3.771081    0.9944329  0.228019301
#> 10      4.974111    3.860621    0.9712608  0.084306583
#> 11      5.415536    3.846274    1.0342747  0.115729752
#> 12      5.473258    3.506657    1.3113371  0.391144916
#> 13      4.712138    3.389201    1.3419927  0.388783864
#> 14      4.156724    3.103787    1.2241299  0.242769625
#> 15      5.260850    3.300805    1.6991991  0.176924401
#> 16      4.439301    3.114704    1.3179546  0.064838406
#> 17      5.005798    3.709997    1.2224186  0.279473695
#> 18      5.741445    3.796873    2.1443245  0.748230404
#> 19      5.006031    3.631104    1.3948293  0.596118848
#> 20      4.810939    3.608596    1.1956757  0.378848491
#> 21      5.964257    3.924781    2.1971734  0.302138935
#> 22      4.666742    2.865453    1.5755346 -0.189670762
#> 23      4.409090    3.750227    0.5967952 -0.081912092
#> 24      4.923912    2.979433    1.3516832  0.255491374
#> 25      4.030662    2.830674    0.7796195 -0.052785959
#> 26      4.256053    2.976168    0.5337846 -0.029980431
#> 27      4.670896    3.722247    1.1050087  0.338284570
#> 28      4.854819    3.695405    1.8163291  0.416973538
#> 29      5.329439    3.835157    1.6383945  0.586389982
#> 30      5.683045    3.257929    1.3965868  0.364857572
#> 31      5.337963    3.432288    1.4017600  0.329455731
#> 32      4.995911    3.656878    1.3146176  0.011042000
#> 33      5.479496    3.457676    1.6532595 -0.004934503
#> 34      5.535347    3.387444    1.7908351  0.460459319
#> 35      5.249392    3.261334    1.4446083  0.077754818
#> 36      5.789727    3.624600    1.7471160  0.599384364
#> 37      5.038649    3.394711    1.0765897 -0.105002223
#> 38      5.171473    3.747639    1.7103717  0.394284976
#> 39      5.355507    3.591811    1.8796752  0.398344015
#> 40      4.521393    3.020589    1.1518016  0.139109754
#> 41      5.443188    3.673622    1.3546794  0.039018969
#> 42      5.711228    3.350665    2.2651181  0.382799545
#> 43      5.425746    3.287375    1.8003016  0.253758380
#> 44      5.702055    3.371482    1.6631095  0.262672053
#> 45      5.291833    3.604040    1.9197447  0.330780946
#> 46      6.200673    3.520373    2.7024526  0.151877874
#> 47      4.572405    2.912330    1.6127664  0.157979240
#> 48      5.187260    3.556976    1.2423134  0.292886543
#> 49      4.997315    3.032280    1.1660997  0.148232628
#> 50      5.577674    3.874288    1.9270803  0.247131092
#> 51      6.393656    2.906656    4.5092467  1.366419555
#> 52      5.641697    2.228838    3.7720595  1.063306470
#> 53      5.840680    2.831340    3.9043326  1.366756791
#> 54      6.485933    3.724917    4.4904473  1.912892518
#> 55      6.183204    2.619041    4.4295409  1.256363041
#> 56      5.814222    2.625688    4.3816317  1.176469845
#> 57      6.392205    2.786451    4.4147181  1.296563399
#> 58      6.771476    2.730362    5.1291759  1.522872891
#> 59      6.271307    3.154903    4.6018426  1.386424961
#> 60      5.563700    3.262149    4.4549919  1.801119084
#> 61      5.454724    2.609713    4.1790310  1.638367657
#> 62      5.821429    2.415635    4.0374044  1.086717712
#> 63      6.005596    3.012737    4.4538477  1.239170988
#> 64      4.351178    2.038772    2.9572945  0.937550106
#> 65      5.669610    2.733455    4.4603550  1.408940010
#> 66      5.906485    2.387384    4.2922137  1.123631912
#> 67      6.213050    3.169951    4.5602313  1.357046531
#> 68      5.543924    2.780519    4.2594041  1.332203825
#> 69      4.914158    2.828243    3.8548661  1.188285032
#> 70      5.937001    2.885602    4.4063282  1.541441079
#> 71      5.809430    3.014567    4.1761300  1.301175300
#> 72      6.737859    3.215718    4.6483806  1.485859800
#> 73      6.440151    2.718641    3.9859716  1.131307129
#> 74      4.764086    2.423966    2.9733789  1.409182565
#> 75      5.501976    2.748479    3.9047333  1.222773911
#> 76      5.757293    2.763216    4.3309152  1.361549201
#> 77      5.949809    2.699480    3.9789540  1.455357141
#> 78      5.296080    2.816553    3.8039383  1.267836736
#> 79      5.959534    3.021541    4.4677717  1.691799056
#> 80      5.787968    2.884260    3.8474397  0.792328711
#> 81      6.074652    3.194213    4.1450577  1.590713822
#> 82      5.115199    2.644884    3.7032588  1.330780514
#> 83      4.792601    2.316472    3.4009651  1.165845855
#> 84      6.076567    2.481070    4.6452645  1.474874127
#> 85      6.608136    2.718490    5.0673787  1.310656325
#> 86      5.462283    2.399104    4.5178435  1.369061917
#> 87      5.713461    2.784873    4.1328858  1.546120323
#> 88      5.576116    2.151503    3.9829453  1.246332541
#> 89      6.092846    2.961205    4.2450705  1.180878293
#> 90      5.191425    2.698978    3.4786419  1.154257956
#> 91      5.123828    2.416957    4.0754817  1.639221981
#> 92      6.520808    2.852264    4.6221344  1.574047459
#> 93      5.569911    3.008814    3.5732518  1.267720562
#> 94      6.217903    3.115587    4.1943995  1.343641750
#> 95      4.347290    2.504060    3.4365969  0.850725974
#> 96      6.642273    3.319005    4.7602279  1.461314268
#> 97      5.438974    2.578323    3.3161728  1.301904763
#> 98      5.419644    2.630886    3.4723436  1.039650356
#> 99      4.871948    2.010969    3.2974710  0.843553186
#> 100     5.821447    2.383579    4.3844550  1.502904576
#> 101     7.641846    3.167858    6.0655229  2.154707236
#> 102     5.901742    2.607830    4.7575524  2.305904559
#> 103     7.229005    3.427091    6.5031063  2.541934468
#> 104     6.313718    3.078350    5.3442407  1.973440057
#> 105     5.872632    2.792882    4.9971586  1.895498670
#> 106     6.058938    2.579562    4.5454694  1.580200713
#> 107     7.396559    3.235222    6.3089493  2.156777639
#> 108     6.537203    3.197767    5.3166212  1.988617827
#> 109     7.726213    3.833256    6.0425999  2.565863904
#> 110     6.931282    3.119668    5.9602116  2.097618999
#> 111     6.432397    2.981308    4.9318081  1.975995760
#> 112     5.927207    2.021215    4.8456643  1.709787922
#> 113     6.194541    2.519301    5.1803503  1.806699500
#> 114     6.415741    2.687137    5.3152899  1.955742199
#> 115     7.053786    2.501241    5.5967865  1.982467484
#> 116     7.760595    3.284177    6.3414294  2.205992258
#> 117     6.476486    2.854868    5.3885570  1.969665494
#> 118     6.556666    2.905521    5.0153401  1.958133937
#> 119     7.050783    3.328746    6.1495920  2.528312705
#> 120     5.858341    2.211191    5.1002442  1.910596177
#> 121     6.570600    3.387790    5.5742150  1.970449792
#> 122     7.349172    3.724642    6.3451077  2.181052462
#> 123     6.721856    2.551526    5.8977072  2.025803802
#> 124     6.489753    3.160693    5.7403409  1.982352998
#> 125     6.866204    3.000369    6.0716603  2.225030499
#> 126     6.256909    2.675722    5.6342411  1.740316318
#> 127     5.979627    3.362984    5.5458099  2.245865059
#> 128     6.260000    3.300863    5.4065481  2.053754324
#> 129     6.362523    2.779490    5.5015961  2.060618049
#> 130     6.999155    3.498574    5.6824508  2.163855923
#> 131     6.123161    2.813041    5.5503558  1.951342425
#> 132     6.867858    3.344665    5.6113283  2.164272856
#> 133     6.897424    3.560707    6.0060216  2.219833932
#> 134     6.685181    3.732176    5.5150194  1.982219494
#> 135     6.330471    2.561527    5.2934664  1.760288338
#> 136     7.576675    3.463090    6.4996977  2.393802747
#> 137     6.925156    3.516107    6.0260803  2.085356391
#> 138     6.637147    3.100464    5.9279226  2.572398076
#> 139     6.103475    2.802750    5.9021550  1.858940683
#> 140     7.521015    2.769759    6.2024016  2.157241086
#> 141     6.452172    2.488301    5.4183062  1.773110194
#> 142     6.242751    3.216069    5.5769470  2.001416239
#> 143     6.305134    2.618899    5.0812712  1.714164001
#> 144     7.322009    3.212439    6.0240137  2.084896351
#> 145     7.025743    2.879282    5.8920155  2.063245199
#> 146     6.588705    3.444138    5.7079175  2.300594098
#> 147     7.345943    2.930963    6.1875406  1.792083455
#> 148     7.136989    2.955429    5.8907362  1.859314596
#> 149     6.305937    2.913291    5.2014702  1.993006699
#> 150     7.239517    3.333447    5.7729348  2.166653615
#> attr(,"seed")
#>   [1]       10403         552 -1184664750 -1005942012  -980626551 -1040701362
#>   [7] -1179563073  -636417366  1093629997 -1499561362 -1775816941   177089283
#>  [13]  -933861518  1169043579  1280489517  1676520102  1660820752   319903727
#>  [19] -1542758422   964640341  1977407510   972534997  -510004074 -1251293649
#>  [25]  -748063227 -2098389969    52893720  -589867052  -967604153  1762497181
#>  [31]  -535087268 -1299474913 -1446149923 -1030279890  -598295276  2005703798
#>  [37]   -28436583  1349124320 -1747465328   534899801  1845409512  1567921758
#>  [43]  1729042415   -88006533   362093993   363155627  1420313748 -1382558940
#>  [49] -1732014781 -1667858230 -1726615486  1001454095  1173408718 -1326018925
#>  [55]  1254824594 -1339157633 -1825260121  -499764968   404486276   112328744
#>  [61] -1370004191    13931685  -385794394  1491207464   899360538 -1783873497
#>  [67]  1172586663  -592391163  -216438406  1719321768 -1628842962   613472134
#>  [73]  -907230181  1998771991  1097149907  -642536244  -248186758  1154822784
#>  [79]  1348767362 -1765242099  1870593067  -641975364 -1547895483 -1752333298
#>  [85]  2124499186  1462470318 -1758195062   407393135   609999457  1158620207
#>  [91] -1304180771 -1907203799  1597182410  1130586258  1595496852   678341881
#>  [97]  1772272208 -1823213943  1345343779  -318429023 -1200192091   -77937468
#> [103]   664595905   771380448   128789495  1754169546 -1776893839  1122264353
#> [109]   828178903  -228965135  1189529516   184679054 -1227695376  -179214412
#> [115]   393696521 -1256924307 -1896933690  -492034197  1357982418  -548564185
#> [121]   480881843   -56603614 -1211250958  -438015150   259693436  -308999273
#> [127]  -967754561 -1398091076 -1119581863 -2034930764  1484519911   559276487
#> [133]   135175446  -705808439  1397941544  1658862133   992672427   235077763
#> [139]   735362254 -1791789309  -359493028 -2104195141   957844273   975138883
#> [145]  -106344563  1852003838  -611298561  1033033532 -1831301125  -565288687
#> [151]  1758441593  -109714187   -85927916 -1723540944 -1839758036  1129361859
#> [157]  1472039005 -1396399207   639592590  1162929198  -492145131  -133671898
#> [163]  1232357788  1505967431  -979830963 -1439085851  -398153620   177754771
#> [169]   725927556  1871601792  -935460052 -1813082847 -1398946387   640491784
#> [175]   513832512  1717538231 -1804808492  -728184900  -291670377  1059716804
#> [181]   633220111  -836192656  -300257037   174748457   -87124704 -1631963151
#> [187]   215448078 -1417691571   768511259  1590123251  1793025346  1134334472
#> [193]    43392204 -1855907695 -1225154811  -635427533   213805693  1286057592
#> [199]  1287735873  -549699488 -1417043918  -159982146  1752442953  -272004212
#> [205]  -987429047   531381104  -945618767  1739910167  1209058761  1806822308
#> [211] -1263545820 -2030879320    97357034  -213915299 -1614828703  1002919562
#> [217]  1963150897  1596692012 -1709103460 -1592895636   777559290  1724377907
#> [223]  1395187239   713725619  -758629034   957163791   435664306  -334208054
#> [229]   103551040  -985233920   462861574   838552733  -742079275  1918182693
#> [235] -1880337861  1416099265  -876499545  -195282115  1770342731  2061110235
#> [241]  -616749339  1793879308 -2115229329  -642012327  -589812932  -794301729
#> [247]   314458221  -588948597 -1858521394  1140639490  1986661460 -1801235148
#> [253]  1638898094  -946155042 -1357800901   -41025887   508366406  1230309974
#> [259]  2037758775  1277689515  -715832793 -1036095965  1020204141 -1290241810
#> [265]  1660847988  -450176230  1783545303   -63571548 -1423916681   884289870
#> [271]  -232738260  -524472730 -1307202449 -1524096456  1234130134  1362189013
#> [277]  1751203335 -1802467632    66574150  1979744341  1601114509 -1821887907
#> [283]  -585329705 -1271856974   366092117  -172760239 -1332667404   300109946
#> [289]  1211118507  1077913377  -529038615  1541456268  1685576952  -911698467
#> [295] -1469508973 -1456977729  -211756076  1203730186  -201301519  1054607221
#> [301]  1874523992     7375572  1557009834   454521088  -331612189   262740151
#> [307]   206686028  1284324531  1393059942 -1046776732   435469805 -1714548034
#> [313]   700349083 -2015255900  1939575287 -2055786081   281266531  1791055453
#> [319] -1949326169  -736156784 -1864329623   399630326  1131229052  -673204053
#> [325]  -322639843    17137754   753172248 -1584554498 -1291245970 -1406839907
#> [331]   -80233061  2089818323  -244431429  1785653838  -469945057 -1942527149
#> [337]  -318440021    78783966 -1447965133  1875292820 -1329670510  -852088942
#> [343]  1004621220   713438242 -1162650773  1319299316  -625424901  1202707155
#> [349]   112473964  1356111635 -1103772738  -761755228   656369594   302040352
#> [355]  1173461899 -1248385159  -188092384  -122954572  1344765316   174041819
#> [361]    20781554  2072616491  1401369736    78358788 -1062664093 -1923062658
#> [367]  2038233820  1125429835  1525460066  -934327172   682971377  -435114735
#> [373]  2128643413  1831524909 -1535337247 -1992294844  2001608449  1968705146
#> [379]  1136036566  2028449828 -1561323116 -1552971097  -864232867  1951454623
#> [385] -1077397679  -417802880   -53892451 -1679909301  1991012594   428632695
#> [391]   605463591   979179204 -2007899632  -431529843  1826015492  -769033314
#> [397] -1014186748  1605074047  1158106154   586472138  1290563807 -1538861551
#> [403] -1180860153   662129636   733878819 -1139636590 -1023802144 -1165492163
#> [409]   772519996  -778367356 -1867183395   851307651 -1410143519  1035320479
#> [415]  -256913808   565568063 -2139208933  -572903543  1342772324 -2003080529
#> [421]  -494598686  1701338939 -1407899697 -1128267200 -1058263393 -1870208535
#> [427] -1360715581  2114004602   656658506 -1856113639  1582063270  1447740264
#> [433]  1016090507 -2009326926  -408831197  1820007982   154999286 -2122939174
#> [439]  -260454809 -1551055145   385880820 -1431394355   837823370  -701800916
#> [445] -1671269756   251469325  1553327402    53016362  1418935681  -770776626
#> [451]  -104079758 -1387175848 -1763385663   870947002  1233133359 -1788742578
#> [457]   958161213  -161769925  -840846445   -76950993 -2060980001 -1455359303
#> [463] -2108172853 -2007660033   476443692  -879205951 -2022139366  -929327727
#> [469] -1490671748  -184420233 -1529957868 -1638633110 -1040458045   551748208
#> [475]   549694960 -1871569517 -2084070686  1453066346 -1011737135  1886129680
#> [481]   156389858  2085851196 -2015092215 -1409551148  2118314879   233774518
#> [487]  1304214550   494073516  2128666496  2096528234  -100152361   344019591
#> [493]  -883081893  1870961290  -695683580  -262525227  -147918288   449599095
#> [499]  2118143562 -1233528816  1634378811  1165682478 -1020053517  1380380579
#> [505]  -370743097 -1550500704 -1136831670   211085709  1355459760  -638416757
#> [511] -1044705684   791637353  1635877927  1022074272  1921461454    74025506
#> [517]  -517968482   256385331  1336264728  -245889011   818858066 -1418803167
#> [523]  -348626277  2076174864 -1113519448 -1902005671   309227900 -1905053105
#> [529] -1914496214  1586419397  1780383352 -1409436750 -1673513864 -1784944219
#> [535]  -804069487  -198021106   437953083 -1122090033 -1358867406 -1266333671
#> [541]   560693859 -1980593625  1371552583   298675323   110964327   120762223
#> [547] -1398019317  1380836559 -1406586177   994051415   203262115 -1212691274
#> [553]  -282018408   388853418   -30342102 -2049413650 -2037953280  1064439531
#> [559]  1971273644   687810913  -625680464 -1943478065   244668045  -975507301
#> [565]  -404088730  -593415771  -173923575  1693305226 -1837575917   473045901
#> [571]  1159833019 -1018302602  1240968534  1498426461  -389951262   418200771
#> [577]  -494598537  1664069855  1011183129   -34845775  1039195468  1188204508
#> [583]   153710037 -1543111843  -695239083   449526944 -1452460317    83424310
#> [589] -1192642873  1042384633   574671913  -452967725 -1392187914 -1556660158
#> [595]  1057113969 -2097235298  -456873053  1845101538   863438154   384717892
#> [601]  1484565545  -228129489   703177001  1704431543  -509582702  -697292547
#> [607]   125200035   146977492   399999509  -334692932 -1965877514   169899279
#> [613]  1501055252  -382788969   668911532 -1259142632   986223206   246401384
#> [619]  1064652553  -127028463  -982662807   784982537 -1272069915   742762582
#> [625]  -639055377   431420803