Simulate Responses from a Multivariate Linear Model

Simulate one or more sets of responses from a Multivariate Linear Model (mlm) object.

Usage

# S3 method for mlm
simulate(object, nsim = 1, seed = NULL, ...)

Arguments

object

a mlm object, typically the result of calling lm where the response is a matrix.

nsim

number of replicate datasets to simulate. If nsim is greater than 1, the output is arranged in a 3D array.

seed

an object specifying if and how the random number generator should be initialized (‘seeded’). Either NULL or an integer that will be used in a call to set.seed before simulating the response vectors. If set, the value is saved as the "seed" attribute of the returned value. The default, NULL will not change the random generator state, and return .Random.seed as the "seed" attribute, see ‘Value’

...

additional optional arguments.

Value

A matrix of simulated values for the response (or an array, for nsim greater than 1)

Details

A simulate function for mlm objects, which simulates one or more sets of responses from a Multivariate Linear Model (mlm) object. If multiple sets of responses are requested, they are returned in a 3D array, with simulation number along the third dimension.

The weights argument is currently ignored -- a constant variance-covariance matrix assumed for mlm.

See also

Author

David Warton <david.warton@unsw.edu.au>

Examples

# fit a mlm to iris data:
data(iris)
iris.mlm=lm(cbind(Sepal.Length,Sepal.Width,Petal.Length,Petal.Width)~Species,data=iris)
# simulate new responses:
simulate(iris.mlm)
#>     Sepal.Length Sepal.Width Petal.Length Petal.Width
#> 1       4.061399    3.132877   0.56605089  0.21101365
#> 2       5.370406    3.424201   1.70697124  0.17904131
#> 3       4.918527    3.013809   1.30413250  0.13444491
#> 4       4.598398    3.396891   0.87473500  0.20891202
#> 5       4.018502    3.204137   0.73667677  0.19509698
#> 6       4.345336    3.171825   1.24269528  0.32546806
#> 7       5.503115    3.923967   1.92831748  0.45030650
#> 8       5.061020    3.318983   1.47275038  0.22239833
#> 9       4.474983    2.937621   1.01764155  0.26344660
#> 10      4.373661    3.000731   0.80665678 -0.18906675
#> 11      4.256926    2.740182   1.14269105 -0.02354808
#> 12      4.299112    2.572156   1.11292586  0.09867379
#> 13      4.034254    2.708025   1.28886722  0.08606510
#> 14      5.692936    3.710207   2.00339627  0.43954085
#> 15      5.335853    3.145755   1.99162277  0.29661860
#> 16      5.287603    3.282925   1.95914143  0.24413527
#> 17      4.398048    3.614692   1.22369261  0.52580236
#> 18      5.406123    3.683904   1.40115167  0.36053712
#> 19      5.496310    3.514280   1.46878005  0.01277441
#> 20      4.468472    3.098875   0.93474528  0.05521668
#> 21      4.486516    3.613056   1.13408753  0.40755360
#> 22      4.789668    3.184635   1.06962778  0.31842073
#> 23      5.284636    3.800429   1.58864296  0.29042940
#> 24      4.977955    3.716080   1.71154308  0.46953204
#> 25      5.457135    3.086960   1.78765989  0.19275725
#> 26      5.189769    3.648046   1.88497972  0.41116940
#> 27      5.065898    3.829154   1.37535061  0.33979249
#> 28      3.893979    3.136767   1.21303057  0.01187448
#> 29      5.135751    3.349462   1.49618917  0.27048171
#> 30      3.708031    2.757481   0.02114788 -0.29608062
#> 31      5.547464    3.279257   1.26191178  0.06426917
#> 32      5.271333    3.573528   1.64501063  0.23724188
#> 33      4.695867    3.609855   1.39255541  0.04329050
#> 34      4.593243    3.540917   1.02736451  0.02957202
#> 35      3.730029    3.073092   0.96547822  0.23279814
#> 36      5.158433    3.685929   1.91757133  0.46715441
#> 37      5.294258    3.567240   1.59404611  0.24169342
#> 38      4.097219    2.711613   0.90956855 -0.12161651
#> 39      4.816296    3.486937   1.51338076  0.33882771
#> 40      4.348486    3.335255   0.77500014  0.12187200
#> 41      4.286712    3.406032   1.30744067  0.24689929
#> 42      5.246896    3.187779   1.57692419 -0.09855883
#> 43      5.432631    3.316151   1.77529570  0.14031378
#> 44      5.865824    3.899126   2.24335043  0.58166996
#> 45      5.694853    3.681827   1.74919233  0.35466063
#> 46      5.681947    3.684566   1.57536912  0.39910738
#> 47      5.369206    3.069702   1.78037762  0.28461054
#> 48      3.962606    3.479889   0.64398050  0.08219230
#> 49      5.034049    3.609892   2.04032017  0.35965355
#> 50      5.246824    4.051571   1.77589019  0.44503090
#> 51      5.311104    2.957471   3.48025773  1.04816998
#> 52      6.470554    3.054990   4.36739717  1.38340030
#> 53      5.922219    3.197902   4.32281291  1.31639208
#> 54      6.149786    2.718072   4.48534505  1.21964200
#> 55      5.486095    3.007289   3.82702246  1.09997468
#> 56      5.648117    2.733262   3.67538944  1.17150119
#> 57      5.729679    2.925901   4.04783190  1.24600736
#> 58      5.363099    2.634803   3.95001031  1.38301860
#> 59      5.679044    2.850396   3.86885526  1.33269686
#> 60      6.910195    2.831583   5.13884692  1.44319302
#> 61      6.055123    2.455951   4.28020430  1.38173304
#> 62      7.256508    3.436067   5.45810352  1.51421326
#> 63      6.767480    2.674096   4.71468580  1.29845731
#> 64      6.135647    3.005570   4.42966523  1.18900203
#> 65      6.240046    3.368902   5.27056902  1.63541043
#> 66      5.483647    2.592188   4.12582496  1.46885584
#> 67      6.424426    3.022260   4.56892828  1.18007354
#> 68      6.151593    2.469163   4.71388367  1.09464664
#> 69      6.395869    2.830305   4.65583973  1.56563440
#> 70      6.488869    3.275332   4.62578806  1.83525814
#> 71      5.803848    2.408167   4.50121696  1.03648022
#> 72      7.179614    3.544105   4.86388343  1.57276253
#> 73      5.888622    2.486689   3.84464297  1.03806047
#> 74      5.962440    2.709144   3.90557898  1.23091161
#> 75      5.288946    2.598119   3.99952315  1.69438880
#> 76      5.898205    2.469267   4.64836179  1.35619774
#> 77      5.661668    2.551190   4.35974362  1.60302960
#> 78      5.858001    2.893289   3.89513521  1.25755217
#> 79      6.882748    2.736665   4.76983434  1.40201389
#> 80      6.344543    2.913025   4.44172238  1.18380880
#> 81      6.579396    2.567562   4.93519383  1.29220591
#> 82      6.278581    2.468046   4.47834190  1.40853643
#> 83      5.872235    2.746481   4.23865100  1.36371603
#> 84      6.290801    3.057846   4.72603735  1.78752689
#> 85      6.305869    2.909807   4.67503918  1.36050013
#> 86      6.150542    2.788012   4.59893202  1.46353496
#> 87      6.374377    2.906300   4.58416470  1.48609929
#> 88      5.841161    2.431206   4.23103020  1.06798727
#> 89      5.032808    2.552570   4.00515139  1.22627552
#> 90      5.295327    2.932499   4.08202055  1.56501938
#> 91      5.435778    2.062900   4.32316518  1.33821403
#> 92      5.938650    2.367012   4.16530169  1.08677448
#> 93      5.361308    2.688134   3.98242731  1.29421113
#> 94      5.608982    2.767787   3.89464052  1.36025183
#> 95      6.143424    2.358864   4.60501140  1.32589973
#> 96      6.074506    2.624730   3.80219220  1.37805244
#> 97      6.745554    2.973685   4.80221503  1.19925318
#> 98      6.272472    2.775673   4.58237981  1.31430929
#> 99      5.024076    2.067920   3.68171006  1.14995480
#> 100     5.894186    2.381162   3.92299584  0.88287785
#> 101     6.495175    3.049690   5.57377990  2.00041344
#> 102     6.800185    2.893239   5.45848249  2.07416687
#> 103     6.127502    3.332351   5.24803711  2.08050483
#> 104     6.670855    3.971155   6.06525098  2.68739372
#> 105     6.310593    2.717941   5.81614124  1.80588671
#> 106     6.350611    2.401894   5.43460451  1.88197223
#> 107     6.083577    2.952487   5.39089669  2.18191893
#> 108     6.417382    2.976001   5.27477226  1.84757171
#> 109     7.560150    3.644692   6.21831383  2.40229346
#> 110     6.645400    3.677645   5.63954585  2.16691600
#> 111     7.519810    3.028540   5.87195160  1.89044780
#> 112     7.422609    3.498744   5.59674088  1.91468216
#> 113     6.962534    3.343343   6.06979650  2.30558919
#> 114     6.916625    2.872828   5.43165161  2.09397742
#> 115     6.193661    2.945526   5.46824749  1.69912335
#> 116     6.483107    3.387858   5.59921870  2.26485513
#> 117     5.981142    2.972996   5.25584787  1.99971223
#> 118     5.828928    2.806493   5.35754722  2.01962370
#> 119     5.172526    2.555604   4.73427464  1.84165541
#> 120     6.454345    2.672250   5.28992843  1.99484714
#> 121     6.311766    2.678873   5.38907731  1.99997115
#> 122     6.890049    2.978604   5.99421733  2.13578169
#> 123     6.763736    2.642747   5.38389947  2.00477234
#> 124     5.350501    2.283314   4.83360146  1.65762881
#> 125     5.892828    2.606021   4.99643756  1.81101407
#> 126     7.357085    3.154430   5.83418397  1.81688041
#> 127     7.041348    3.193595   5.62559792  1.89484811
#> 128     6.765536    3.055026   5.75670342  2.02590541
#> 129     5.958641    2.252256   5.29389269  1.88765775
#> 130     5.706778    2.890928   5.01027413  1.94834029
#> 131     6.551741    2.472030   5.51508204  1.79381124
#> 132     7.316853    3.528538   6.55763053  2.28409591
#> 133     6.188417    2.097755   5.50283907  2.01220035
#> 134     7.024171    2.829472   5.55594943  1.78362500
#> 135     6.021206    2.433782   5.25107765  1.77923566
#> 136     5.930597    2.683401   5.40414614  1.85872355
#> 137     7.086997    2.936349   6.09346391  1.97227227
#> 138     6.118796    2.937390   5.31238390  1.62573784
#> 139     6.487870    2.947962   5.06264649  2.16562950
#> 140     5.946847    3.375034   5.36243200  2.22359847
#> 141     6.537015    2.847637   5.10308071  1.73117367
#> 142     6.815713    3.535073   5.79348820  2.50568735
#> 143     6.177659    2.526300   5.52663118  1.82853649
#> 144     6.480701    3.195966   5.27959067  1.77032572
#> 145     7.407407    3.102607   6.17007763  2.16675410
#> 146     6.102596    2.509696   5.60523807  2.09417171
#> 147     6.109878    2.430809   5.25715053  1.74618490
#> 148     6.842517    3.243259   5.81170775  2.14217389
#> 149     6.607946    2.670433   5.85118130  1.89489182
#> 150     6.587880    2.783570   5.86143165  2.12125431
#> attr(,"seed")
#>   [1]       10403         383 -1997980237   594856901  1342562616 -1627096546
#>   [7]  1390911484 -1894299704  -567575123 -2143251702   768518711 -1947623505
#>  [13]  -944535052    91800155   906762345  1462414608   213660260 -1252776625
#>  [19] -1886650976   486704093  -664658783 -1827214704   624933425  -962425633
#>  [25]  -741406872  -987081722  1437739060  -542570710  1987565203 -1272695064
#>  [31]  -861054824 -1147247007  2104059511  -214016380  1061886562 -1520452946
#>  [37]  2092136781   863827118 -1188640672  -741372692   828202383 -1374219348
#>  [43]  1254816941 -2122100951  1324012411  1432798237 -1312901496  -407653607
#>  [49]  -158723269  -874930662 -1996684487  1408982129   397779028    67046845
#>  [55]  1071051081 -1209566802  -172096692 -1814109766    69117589  2031832841
#>  [61]   804468815 -1325541750 -1798464076  -911083671   975326842  -605256465
#>  [67]   824734486   908987330  -635358356   104979110  -855377426 -1058651118
#>  [73] -1623509035  -195562377  1886740839  1030219450 -1623310279  1841087582
#>  [79]  -990489253  1064178246  -947781550  1942815280  1690794496 -1985962054
#>  [85] -1057226950  -820000749  -674963571   696463878  -781424551   313884492
#>  [91]  1730586169   794752189 -1739416073   319463423  -153259732 -1631812903
#>  [97]  -255558337 -1080477392 -1549912670  1556320553  -338232733   379636688
#> [103]   935372938  -654738803  -875802854  1221850462 -1579285962   755636072
#> [109]  -911530558  -473964218   784191618  1386106438     8260280 -1395447545
#> [115]  1125680010   762009819  2044065726 -1232007073    45219469   247879190
#> [121] -1816532370   289410208   723630817  -559588969  1079208253  1160388729
#> [127]   564379019  1949057167 -2075019539   601643991   786037468  1347654433
#> [133]   141656280  1609211012 -1661158011   683796054   483521445  1729172030
#> [139]  -982563508  1144943894  1203689116  -448721087  -180939779 -1097165204
#> [145]  1946512176 -1093470290 -2058335748  -589206861  1320930962  -832254772
#> [151]   952112190   609050093 -1726485453  1561208656  1766586252  -919389997
#> [157]  1982175831   732794808  -915861838 -1799592778  1273263647  1913257608
#> [163] -1853987092   476003894 -1671602873  -635787713   328386073  -300212427
#> [169] -1354955518 -1471283685  1470328265  1446161643 -1158570225 -2112768494
#> [175]   540770924  -790138465  1290799369  -711190561   812740415  2132905851
#> [181]   255030446  -222305319   587847589  -598483055 -1716862074   363386291
#> [187]  1948624082  1198883755 -1967537875 -1553222411  1776878614 -1783669255
#> [193] -1162286625  -883251058  -816217783  1277934019   315984326 -1267480773
#> [199] -1272686917  -555490958 -1019725216 -1374581474 -1729369557  -155308343
#> [205]  1220376886  -222765040  2043564979  1463905933 -2038862846    45801306
#> [211]  -334182153 -1059008713 -1061141338 -2009658170  1068949225    42734807
#> [217]  1891631684  -923352760   661845776 -1349022906  -408431758  -576625935
#> [223]  1067951152   654716591   313082746  -724651842  -990314991  1490869503
#> [229]   821474674  -583493963   703492705  -208906177  1236898904 -1992832752
#> [235]  -598322501  -603241358    90941545  -291107746  1224520197 -2009917122
#> [241]  1683698210  -526680439  -659237644  -642273127    79693476   220069730
#> [247]   -74527397  -386733455  -974600136  1954593898 -1813930716 -1915156645
#> [253]  1140654025  1510683522  1625430695   126866420    84211853  -227643088
#> [259] -2060643655  -501808232 -1734487309 -1756256016  1215421807  1799345814
#> [265]  -246510819  1529215268 -2019409797  -714914787 -1675771785 -1226935445
#> [271]   680275216  -883518979  1062291230   515086058  -815422110 -1732282957
#> [277] -1827956073   272642475  1503450654 -1363192534  1602859116  1927374837
#> [283]   256947126   266649659    43775859   622974481  -353809860   118742769
#> [289]  1307072866  1479574641  -674607853    60948782 -1764034097 -1882727161
#> [295]   400991597  -308499252   176684789  -548509773    23062617 -1516884317
#> [301]   185383101 -1397263080   780779784 -1767663087  -662008188 -1115494850
#> [307]   502022509  1890168628   202512801 -2145359302 -1739102302  1829957766
#> [313] -1283613886 -1727331496   252007001   -21631555  1567283526  1666025829
#> [319] -1663134466 -2080010463   233052260  -526131490  -146151602  -514371807
#> [325] -1970312691 -1304886417  -708041204   537533225  1601023281  1329693090
#> [331]   969008454 -1068976209  -169852616   373486091  2001372974 -1425025079
#> [337]  -598403861  -452099160  1583834401  1545760174 -1717978362  1981198934
#> [343]  -124253771  -432920304  -253861143  -444732842  1862189174   894333550
#> [349]  -851872087  -748690879  1321354704   751368396  -255737936  1673439028
#> [355]   182180367 -1279687249  -426095694  -346428839  1364832551  1576662233
#> [361]   990459487  -931998615 -2076668985  1943408554   972912204 -1621415496
#> [367]    28236987  -481179414  -284209388 -1820019270   363613814 -1341198624
#> [373]  -467771830  -729935214 -1277079053  1692480859   114197806  -839754393
#> [379] -1772720812  -186803823  2073776088   813321452 -1791490580  -527854165
#> [385]  -711590749  -813754621  -818758912  -521730277   238579619 -1954462659
#> [391] -1643894706  -334515641 -1275554964  1949664007 -1546459341  -752588270
#> [397]   690682886 -1134042793  1660590437   921043601  -324430869  1772615930
#> [403]     1905431  1426658567  1135823509  -174268504   941523853  1621113626
#> [409]  1479975495  -783661180 -1007930123 -1851965465  -393044396 -1451943645
#> [415]   215943242   632876094  1058677345 -1975239351 -1830492816   387806725
#> [421] -2021415577  1771616024   407092449   396671143   483299476   636873395
#> [427]  1820047252 -1087673405 -1879971935 -2075209464  1250453853 -1822164659
#> [433] -1563629866 -1075780514 -1023794941  -321286548  -680400366  1686352265
#> [439]  1535886131   971572801 -1839205531   453595454 -1959095499  1773540952
#> [445] -2033304873   -95392242   792245119  1612868031   171869523  1463809476
#> [451] -1148826097  -356594669  1960191941  -529010617 -1361137735   984388801
#> [457]   741085538  1720637762  -320888560 -1856095739 -1675369321  1650239280
#> [463]  1364238648  1700599569   809582938  -200338167  1842884386 -1263822600
#> [469]  -695311606  -659452147 -1856861670   -54374257   -80307871  2083067151
#> [475]   238093500  1213317625    16440340    51498749  1296941106   966857441
#> [481] -1406359340 -1338822991 -2037004846 -1437586100 -1945545957  -352031875
#> [487] -1490278542  -553851110   712528012  -961366030  1084038213  -239221677
#> [493]  1912776616  1379890837   544173987  1058531760   437050664   272441899
#> [499]  -783958467    52897280  2058810630  -904579095 -1667881670 -1546688429
#> [505]  1353475321  -534187151 -1526879453 -2110824397  1645876375 -2025756674
#> [511]  -497957251   367338051   873911818  -711718984  2092903818  -271791336
#> [517]  -141315279  -862658068   -60880479 -1405937550 -1655407979 -1088785427
#> [523]  -578966419 -1316657293   447582963 -1004060195   446684468 -1909079483
#> [529]  -899364805  -980684797 -1202497424  1887151859  1975034472  1528720261
#> [535]  1290055920  -217256835   183434994  -118742925  1453058900 -1712478326
#> [541] -2095714022 -1567684605 -2003791191   547399854  1693943960 -1933927676
#> [547]   -41822988  1003593015  1882185768 -1737298734 -1982289978  -316112766
#> [553]   387944111   785953906  2038423075  -914631304 -2067381565  1469402576
#> [559]  2145269418  -711994289  1272305470  1501269432  1573805380   902698351
#> [565]  2004546165     9376758 -1514690817 -1655348840  1326507831  1661036525
#> [571]  1924085427   -22814757 -1648537057  1403943384     8918089  -326372852
#> [577]   207899434   737404908  1752838118  2123580875  1678017529  1222682508
#> [583]  -603133119  -984306896 -1522755690 -1074188333  -920254685  1663995404
#> [589]  1110921320   -86689537  1845981081   639205562   557196689   316333783
#> [595]  2070819499   169250521 -1942144866 -1865562054   -23994413  2026685099
#> [601]   115548010  -438550086 -1416430154  1496841221  1632935541  1022124783
#> [607] -1201815147  -235878681  1606894697 -1517704076 -1127576756  1843972959
#> [613]   274171407  -179252320  -600258644 -1266376361  -510113365  -356978791
#> [619]  1653203598   616991215 -1581281392  -734458249  -904068485  -470644881
#> [625]   833360031    -7948445