Simulate Responses from a Multivariate Linear Model

Simulate one or more sets of responses from a Multivariate Linear Model (mlm) object.

Usage

# S3 method for mlm
simulate(object, nsim = 1, seed = NULL, ...)

Arguments

object

a mlm object, typically the result of calling lm where the response is a matrix.

nsim

number of replicate datasets to simulate. If nsim is greater than 1, the output is arranged in a 3D array.

seed

an object specifying if and how the random number generator should be initialized (‘seeded’). Either NULL or an integer that will be used in a call to set.seed before simulating the response vectors. If set, the value is saved as the "seed" attribute of the returned value. The default, NULL will not change the random generator state, and return .Random.seed as the "seed" attribute, see ‘Value’

...

additional optional arguments.

Value

A matrix of simulated values for the response (or an array, for nsim greater than 1)

Details

A simulate function for mlm objects, which simulates one or more sets of responses from a Multivariate Linear Model (mlm) object. If multiple sets of responses are requested, they are returned in a 3D array, with simulation number along the third dimension.

The weights argument is currently ignored -- a constant variance-covariance matrix assumed for mlm.

See also

Author

David Warton <david.warton@unsw.edu.au>

Examples

# fit a mlm to iris data:
data(iris)
iris.mlm=lm(cbind(Sepal.Length,Sepal.Width,Petal.Length,Petal.Width)~Species,data=iris)
# simulate new responses:
simulate(iris.mlm)
#>     Sepal.Length Sepal.Width Petal.Length   Petal.Width
#> 1       4.696832    3.572376    1.3041520  0.1511658447
#> 2       5.339344    3.540815    1.7034210  0.2428924962
#> 3       4.505305    3.124200    1.1917156 -0.0824897943
#> 4       5.137981    3.584217    1.6265706  0.2638162417
#> 5       5.752638    3.918486    1.8894249  0.0946923192
#> 6       4.826254    3.264171    1.4561954  0.1760241367
#> 7       5.431103    3.600542    1.4801647  0.2475318981
#> 8       4.631923    3.130575    1.3660198  0.1335249856
#> 9       4.656604    3.366932    1.1877276  0.2773946061
#> 10      4.665275    3.441435    1.2645257  0.0027430141
#> 11      5.299351    3.599160    1.9843233  0.2927272781
#> 12      5.895697    3.724344    1.8166630  0.1091625247
#> 13      4.479201    3.218013    0.9835639 -0.2140429228
#> 14      6.023320    2.944452    2.2966644  0.2992804774
#> 15      4.320908    3.190697    1.0866110  0.4287502623
#> 16      5.122727    3.898800    1.0942278  0.1124004716
#> 17      4.766134    3.241762    1.5249288  0.2505142489
#> 18      5.298947    4.287984    1.3347266  0.1730475822
#> 19      5.118469    3.546170    1.5656846  0.2677512338
#> 20      4.982124    3.510732    0.6510391 -0.0227163895
#> 21      6.029839    3.902253    1.8295413  0.4726555217
#> 22      5.843102    4.457883    1.5000211  0.1423915539
#> 23      4.437257    3.689596    1.3143166  0.2649296761
#> 24      4.719672    3.634443    1.0153420  0.1120165556
#> 25      4.909707    3.392966    1.5482204  0.3500334258
#> 26      5.060643    3.810759    1.0019619 -0.0645106289
#> 27      5.540132    3.086225    2.0287025 -0.0695682222
#> 28      4.854425    3.549900    1.1720488  0.1987510008
#> 29      4.946349    3.324432    1.4944836  0.5220421242
#> 30      4.560641    3.171673    1.0205289  0.1337780423
#> 31      5.149805    2.928370    1.9546713  0.1922517178
#> 32      5.097000    3.390392    1.4584913 -0.1841899724
#> 33      4.783529    3.852442    1.2012059  0.2435083328
#> 34      5.131552    3.407811    1.7248558  0.2376329837
#> 35      4.009331    2.595429    0.3121248 -0.1582477506
#> 36      4.224501    3.154224    0.8928307 -0.0566796122
#> 37      4.911608    3.263305    1.6863192  0.2249610697
#> 38      5.515692    3.574295    1.7604806 -0.1219340566
#> 39      4.302295    3.288824    0.7978758 -0.0969257578
#> 40      5.253488    4.200156    1.6198926  0.3577452623
#> 41      4.646163    3.363022    0.4717280  0.0004231092
#> 42      5.173576    3.718440    1.6205488  0.7473822851
#> 43      4.382395    3.593502    0.9829820  0.1244283540
#> 44      4.968843    3.446804    1.3303077  0.3788285728
#> 45      5.538912    3.930446    1.9307695  0.4223042283
#> 46      4.900654    3.451272    1.3868461  0.3111903195
#> 47      4.371102    3.056201    0.7332139 -0.0177216494
#> 48      5.619583    3.989164    2.1298399  0.5229677613
#> 49      4.987682    3.406652    1.1288149  0.1387578232
#> 50      5.257762    3.454967    1.3542922  0.3216034672
#> 51      5.368920    2.542354    3.7356475  1.2081923170
#> 52      5.656966    2.720760    4.2524704  1.5884578829
#> 53      6.540001    2.161826    4.4967917  1.2013437420
#> 54      5.345647    2.110818    3.6816238  1.1094045696
#> 55      5.378997    2.990971    4.0302473  1.4151627325
#> 56      5.784599    3.198468    3.9795863  1.5295336888
#> 57      6.044785    3.151506    4.3011960  1.4599936737
#> 58      5.128732    2.651998    3.9989967  1.4158442882
#> 59      6.702790    3.262130    5.5209655  1.5459558152
#> 60      4.874711    2.480785    3.6956862  1.2257207621
#> 61      6.324677    3.011754    4.4627574  1.2075695874
#> 62      6.549424    2.937380    4.6461847  1.3119391947
#> 63      5.528461    2.813929    3.6503182  1.4212720646
#> 64      5.236189    2.296087    4.2531095  1.5320335976
#> 65      4.882663    2.557257    3.5100234  1.2019082628
#> 66      6.170815    2.951437    3.9323796  1.4653143745
#> 67      6.394728    2.826838    4.1704162  1.4262403034
#> 68      6.121575    2.679648    4.7576101  1.1844714385
#> 69      6.234670    2.770271    4.1116180  1.2714379900
#> 70      6.609913    2.912316    4.7929196  1.6216235017
#> 71      6.467939    2.913363    4.6405591  1.5316685712
#> 72      6.388553    2.768654    4.2323421  1.7203952756
#> 73      6.022505    3.250747    3.9302597  1.3181673001
#> 74      6.655209    2.857117    4.0370812  1.1309789048
#> 75      6.098546    2.710736    3.8563479  1.2614821244
#> 76      5.214521    2.531558    3.8544476  1.5730658846
#> 77      6.511141    3.089381    4.5959347  1.4523953146
#> 78      6.416399    2.927332    4.3449343  1.3013194882
#> 79      6.318162    3.189821    4.5959736  1.4885363999
#> 80      5.313453    2.426748    3.9151969  1.5961254077
#> 81      5.899160    3.009207    4.7034256  1.4531230672
#> 82      5.978372    3.093194    4.4357192  1.4636200310
#> 83      5.176239    2.744875    4.2692373  1.3021218909
#> 84      5.676488    3.161060    4.2034905  1.3143628842
#> 85      5.744784    2.608641    4.1553900  1.1892984590
#> 86      5.504455    2.580926    3.4669527  1.1790638285
#> 87      6.250794    3.028190    4.4538344  1.3097291656
#> 88      6.146755    2.669971    4.2657843  1.2277001120
#> 89      5.983681    2.890883    3.8335333  1.1011663393
#> 90      6.052876    2.598355    4.0770474  0.9834550295
#> 91      5.517866    2.725351    3.3757483  1.2057981076
#> 92      5.898951    2.701091    4.2263197  1.2981074843
#> 93      5.845676    2.785234    3.9312640  1.1814409891
#> 94      5.200826    2.362377    3.8575777  1.4998154822
#> 95      6.196641    3.127417    4.6007966  1.6285278211
#> 96      5.205395    2.295457    4.0504537  1.3281829693
#> 97      6.995195    3.130206    4.7922647  1.8000820463
#> 98      6.086890    2.736249    4.4065203  1.7339306628
#> 99      7.284807    3.403689    5.4157839  1.6927845356
#> 100     5.666785    2.366249    4.0565440  1.1114712126
#> 101     6.329513    2.772929    5.4599736  1.7383406249
#> 102     6.371631    2.987411    5.0541158  2.0011391656
#> 103     6.573878    3.317188    5.8976695  1.8911401340
#> 104     7.243546    2.581586    6.0087514  2.1899689451
#> 105     5.898398    3.078192    5.2331547  2.2412064717
#> 106     7.679513    3.552691    6.8089190  2.6194844723
#> 107     6.330956    3.023179    4.9778697  1.8866794294
#> 108     6.242913    2.997848    5.6758760  2.2384206074
#> 109     5.376875    1.790039    5.0324839  1.9951488265
#> 110     6.894602    3.122095    5.7581341  2.0955190072
#> 111     6.815766    1.995218    5.4418224  1.5770168319
#> 112     6.299483    2.351457    5.6027632  1.9111550272
#> 113     6.461937    3.421714    5.2927482  2.1223777957
#> 114     6.784220    3.286198    5.7328027  1.8197602272
#> 115     6.743088    2.955457    5.2575978  1.8073005034
#> 116     5.469414    2.062692    5.4529218  1.7393427267
#> 117     6.907914    3.239858    6.2527411  2.4081567722
#> 118     7.086962    3.317193    6.0326050  2.2976181479
#> 119     6.638805    2.538085    5.5179960  2.0842067527
#> 120     7.216819    3.718865    6.2922809  2.4425678637
#> 121     6.907476    3.630146    5.3646751  2.5652723805
#> 122     6.533227    2.254031    5.9233052  1.7761248496
#> 123     6.128936    3.125595    4.8579039  1.8200741919
#> 124     5.898300    2.852970    5.3327547  2.1427197198
#> 125     6.703015    2.660149    6.2474538  2.0894078465
#> 126     6.776989    2.942562    5.5574981  1.9500561502
#> 127     6.466723    3.504093    5.2841434  2.1657699082
#> 128     6.767649    2.878887    5.3709213  1.9446954792
#> 129     6.610375    2.750731    5.6211589  2.0584063335
#> 130     7.053469    3.045491    6.1274207  2.0986316236
#> 131     6.326071    2.860573    5.4169723  2.3338929674
#> 132     6.690484    2.735308    5.8031495  1.8363035582
#> 133     6.561478    2.733763    5.5104557  2.0104039499
#> 134     6.333314    2.318161    5.0306031  1.8759437972
#> 135     6.296285    2.812689    5.2648615  1.9646261922
#> 136     7.212238    3.141477    6.3265427  2.0929076887
#> 137     6.820244    3.220766    5.4889302  2.1043229777
#> 138     7.007908    3.171120    5.8219648  2.1256456671
#> 139     6.720647    3.025704    5.7057748  2.0717978032
#> 140     7.206176    3.067800    6.3106839  2.1406030273
#> 141     6.939192    3.079563    6.1333377  2.1970547501
#> 142     6.174569    3.198252    5.5050227  2.1863728537
#> 143     6.583207    2.713196    5.3090710  2.0189852122
#> 144     7.200422    3.389015    6.1469269  2.2615273697
#> 145     6.999490    2.859268    5.9869491  1.9294372900
#> 146     6.709311    3.017308    5.6140944  1.9404386044
#> 147     6.377161    2.460282    5.4754440  1.8520804455
#> 148     8.320805    3.545339    6.5606588  2.2624594907
#> 149     7.399217    3.142827    5.4224632  1.9965228480
#> 150     6.252032    2.870578    5.3919467  1.6493308136
#> attr(,"seed")
#>   [1]       10403         190  1547775040 -1574186431 -1641968118   803269845
#>   [7] -2129081757  1837318449  1510610021  -470513295 -1813945152 -1533463939
#>  [13] -2125581043  1771176773 -2042825169   840619493   135716594 -1251537652
#>  [19]  1637589536   239590300  -548282318 -1130688563   704007949  -587716839
#>  [25] -1708181123  1267990004 -1296756687  -967160741  1850877284   987456510
#>  [31]  -154175130  1336181234  -959553111   850672696  -951522152  2134163904
#>  [37]  -763743781  -129553845 -1701483728  1134975944   315220446  -231713727
#>  [43]  1062199823  1990150483   310592772 -2014770110  1414032159  1328797725
#>  [49]  -198160066  1790074822  -287148567  1463160275  -510081625   729396423
#>  [55]   406889189  -563037082  1136301339    45328309 -1979239218 -1298517253
#>  [61]   844609494  -591628648  1182720355  1635963405 -1649434234  -161942295
#>  [67]   487569041 -1667655054 -1414793521 -2134708717  1757569445    15677857
#>  [73] -1384430817  1640398218 -1591780947  1147410908 -1774856359 -1336039128
#>  [79]   137806125 -1124639013 -1448268633  1805095974 -1614276000  -417073096
#>  [85] -1553730573  -385626088   205953440  -944012116  -565245695   571635592
#>  [91] -1602213001  1012976868  1794134686 -1718167105   932891334  -368109701
#>  [97]  1717372234  -880145132 -1903874242  1230295930  -209728355  1794683198
#> [103] -1636589940 -1131200015 -1405058657    95764250   606032867  -136003268
#> [109]   303467105 -1349402743 -1496852335    13350512  -950702089   -48574332
#> [115]    71985415   998253786  -160007661 -1469578398 -1901025019  1419983603
#> [121] -2128409901   -77361692 -2030535438  1307360002   375610229  -907388659
#> [127] -1412789081    68419966   -37078705   769698904  -439560455   705381460
#> [133]  -594491205   139123297 -1977681397   913622696  2010293447   772739054
#> [139]  -214109618   457286202 -1008523718 -1756867732  -943428494 -1902229022
#> [145]  2045659001  1683343636 -1298726075  1621155210  1610401433   578049299
#> [151]  -820221441  1429278124 -1923450752   -28640910 -1425444556  2103368592
#> [157]   641705816 -1130658643  -611498761  1085350801  1491657604   330474060
#> [163] -1095624381  -870876046  -915144577 -1180894647 -1274918642  1774233070
#> [169]  -389934183  1239074643   666668212   330341032 -1896885008  1152824573
#> [175] -2053945355 -1889377400  1590271676   120413137   656734529   374896167
#> [181]  -324387566   655410050  1040587591 -1457724076   -18544666  -497487829
#> [187]  -118717125  1629609399  -946638959 -1209371940  -778183188  1837217740
#> [193]   729447524 -2039104377 -2042606501   877658546  -174288150   325641597
#> [199]  -176778391    45453632   861274990 -1102081361   207934019 -1455384102
#> [205] -1679287291 -1111338673 -1335103233   -72171965  2029243070  -294476602
#> [211]   308202916   757832078 -1208487136  1644262915 -1444471243   820251542
#> [217]  1657780389  -713595927 -1027301486   460114259 -2040735962   -22476606
#> [223]  -850535235  1083133522  2003625856   619422833  1588518722  1577500207
#> [229] -1444776709  -806828611  -947239303  -951040134   800998727   622341511
#> [235] -2023221070  -277944378   213182987  1244003060 -1446918011  1690233623
#> [241]  1627089176  1117847238 -1114051777  -907359716   642801331  -830666027
#> [247]  -335724301  1561218472  1359461431   582801049    95018757   654921829
#> [253]  1232723118  -881692722 -2063771639  1338670514  -201454454 -1375641301
#> [259]  -652518910  1464986898   817673819  -505769177 -1162261492  1698980627
#> [265]  -669199199  1714994826  -927235662  -505517336 -1383016390  1899965518
#> [271] -1570014237 -1548426327  1832814316  1625346634 -1000572441  1116358593
#> [277]  -198129997  1828647568 -1476206199 -1577645334 -1273863534  -631071289
#> [283]   -53784348  -668455648 -2103180474  1906888573  -932150717 -1452501306
#> [289]  -758522198 -1001903052   726234864   241592242  1208433547  2132944308
#> [295] -1280957226 -1587934021   753821229  2041620333   913342837  1117107325
#> [301]  -309144618   649714507   891082689   432544780   928628554  -514776947
#> [307]  -121738528  1126530097   622806970  -722181913  -767551981  -591591920
#> [313] -1801398759  2082842872   238713342  -341811218  -836480101  1080132123
#> [319] -1164153185   547757208  -417929646   540392521  -795278324  1528962492
#> [325] -1853815627  1998344872   836782925   -19541431   559429388   872571860
#> [331]  1342636317 -1986368274  1397222239   641691129  1056819094  -781890184
#> [337]  1112248196   332625296   346945453   696351882 -1256464617  1892918870
#> [343]   344593632 -1741735364  1967954727 -1656416355   444357126 -2085301361
#> [349]   558715870  1069370599 -2012449088 -1860347883 -1921237909   553099792
#> [355]  -567169844   604576432  1005635878 -1307866292   901406587 -2117586419
#> [361]   615586571  1859414567   990635111  -403237314  1802690081 -1898754152
#> [367]   840061576 -1911236582 -1855317906  -109187984 -1100528251  1855548446
#> [373]   543373220 -1116543478  -336512100    20532020   318875603  1630914182
#> [379]   556847726   225986919   291993047  -212416569 -1222494626   154591880
#> [385] -2027020754   181518345  1070214861   920020183 -1026251531   876415001
#> [391] -1677087326   256955964  -595896762  -100822960   771578925  1015557755
#> [397] -1189382341  2002793642   684140614 -1007665551  2106441954   926144158
#> [403]  1753189393 -1384056884   -39397314   930073170   808480040   260239171
#> [409]  -707728854  -322774661 -1693296481 -1832340005  -319926209 -1110320732
#> [415]  -782361477  -656729987 -1256702410    -6649208  -168237990 -1541475174
#> [421] -1931345189  -781574447 -2109802909   363834819 -2050877428  1167293104
#> [427]  1442569710  -455114395 -1501840742   819568607  1202034848  -247650733
#> [433]   469189565 -1086622620  -562158782  2042277947 -1018353752  1179238971
#> [439] -1592490043 -1530837173  -250771170 -1812911109  1948195194 -1131542669
#> [445]   979260506  -385158425   460071024  -377724766   210175164   879746910
#> [451] -1938071227 -2136894144 -1619827157  -637964730   152881652 -1007831687
#> [457]   -75969120   345004243 -1672729327  -527714732  1800996400   218592704
#> [463]  1014053965 -1258714664   233620137 -1708101306  -976462802  -628926507
#> [469] -1436667514 -1834046974  -939909093  -879929236  2089642200 -2128945805
#> [475]  2086262471  -709224722  -232301258 -1870293364  1518568081   -23246200
#> [481] -1552375778   944538612   230311678   336121621   -62797527  -575649504
#> [487] -2133079426 -1106727909 -1872718536  1298205365  -819583779  -822732257
#> [493]    75984658 -1492007527 -1707198458 -1606581406  -143517715  1335836663
#> [499]  -475556881  1953662496  1646194609   936183101  2010208410  -601046595
#> [505]   281710310  -353525191 -1891868883  1000945986 -1607446406 -1374484409
#> [511]  1423084750  1103090342  -538961121   108722883  1275059379  -679647499
#> [517]  1753982392   814367483   697863877 -1856601093   304286936 -1535726172
#> [523]   767659268  1674354974    24246122 -1174024387   446482180  -904354931
#> [529]   225103755   222766808  1757816267 -1645044974  1574064181  -353693667
#> [535]  1833730588  1272732627 -1052540816 -2035187939  1157276583   458083048
#> [541] -1600787040   211553393  -927908589  1091609923   383630988  1554627431
#> [547]  -405013406  1212707977   448458580  -771523577 -1967584811 -1993350882
#> [553]  -515232175   801719997   575765151  1110255925  -748418401  1061180215
#> [559] -1725007739    19279744   350363933 -1572502753 -1262676944  1974744986
#> [565]  1717718174 -1650508943    52293274  1182976485   109418118 -1411145611
#> [571]   404805183   585072732   705887081   604186047  -450529469  1548190941
#> [577] -1828961508  1133590033   314388072 -1867476838   551292562 -1167747589
#> [583]  1391612904 -1773963443   942725900   247234817  -542400475   998511708
#> [589]  1069731409  -554742864  1929811310  -123801912 -1523406037   390310521
#> [595] -1296579688   992616323 -1711989463   720623414   424333095  2004602718
#> [601] -1900620799 -1197535611 -1596720960  1740402000  1783152236 -1957306190
#> [607]  1585489222 -1831617375 -1260623334 -1782610825  1068682169 -1496104994
#> [613]   596381612 -2092981650  -327117732 -1883244078   586477920   209210572
#> [619]  -768226850   147795577   546716642  1739159752  -172034238   144472719
#> [625]   471442994   115850854