Simulate Responses from a Multivariate Linear Model

Simulate one or more sets of responses from a Multivariate Linear Model (mlm) object.

Usage

# S3 method for mlm
simulate(object, nsim = 1, seed = NULL, ...)

Arguments

object

a mlm object, typically the result of calling lm where the response is a matrix.

nsim

number of replicate datasets to simulate. If nsim is greater than 1, the output is arranged in a 3D array.

seed

an object specifying if and how the random number generator should be initialized (‘seeded’). Either NULL or an integer that will be used in a call to set.seed before simulating the response vectors. If set, the value is saved as the "seed" attribute of the returned value. The default, NULL will not change the random generator state, and return .Random.seed as the "seed" attribute, see ‘Value’

...

additional optional arguments.

Value

A matrix of simulated values for the response (or an array, for nsim greater than 1)

Details

A simulate function for mlm objects, which simulates one or more sets of responses from a Multivariate Linear Model (mlm) object. If multiple sets of responses are requested, they are returned in a 3D array, with simulation number along the third dimension.

The weights argument is currently ignored -- a constant variance-covariance matrix assumed for mlm.

See also

Author

David Warton <david.warton@unsw.edu.au>

Examples

# fit a mlm to iris data:
data(iris)
iris.mlm=lm(cbind(Sepal.Length,Sepal.Width,Petal.Length,Petal.Width)~Species,data=iris)
# simulate new responses:
simulate(iris.mlm)
#>     Sepal.Length Sepal.Width Petal.Length  Petal.Width
#> 1       5.474295    3.500827    1.5006556  0.107795130
#> 2       3.860236    2.890695    0.9760483  0.143557237
#> 3       4.620162    3.259003    1.1673980  0.002933171
#> 4       4.880304    3.596769    1.4279064  0.366431909
#> 5       4.808359    3.420374    1.1584776  0.275034160
#> 6       5.133711    3.014254    1.5877389 -0.120020326
#> 7       5.307697    3.706264    1.4434289 -0.046449738
#> 8       5.225727    3.601139    1.6759869  0.451182168
#> 9       4.196648    3.181693    1.4125789  0.294099558
#> 10      4.701931    3.313628    0.9838511 -0.015428931
#> 11      4.810880    2.876809    1.2705556  0.107402855
#> 12      5.649296    3.126024    2.0362208  0.506682261
#> 13      4.890013    3.050889    1.4102130  0.044635643
#> 14      4.586297    3.467854    0.9745025 -0.015875631
#> 15      4.550655    3.290323    1.1603624  0.043827589
#> 16      5.840931    3.624583    2.0909875  0.456323911
#> 17      5.967401    3.704423    1.8430066  0.245335909
#> 18      4.472686    3.290322    1.3106864  0.579728036
#> 19      3.879831    2.740740    0.5702311  0.445212758
#> 20      5.039837    3.083959    1.9313953  0.191990685
#> 21      5.834584    4.535491    1.8729021  0.732779035
#> 22      4.578259    3.659707    1.2993552  0.283706204
#> 23      4.549375    3.183727    0.9468910  0.546029048
#> 24      5.341235    3.350018    1.8892231  0.236570073
#> 25      5.453537    3.812660    1.6488595  0.165738165
#> 26      4.306800    3.554187    1.3093106  0.353192674
#> 27      5.029856    3.876552    1.3510406  0.139571489
#> 28      5.423810    3.818467    1.7230902  0.531040891
#> 29      4.898245    3.933487    1.5755236  0.243260753
#> 30      4.817135    3.483528    1.0220347  0.228797247
#> 31      5.570520    3.541628    1.6058634  0.474481173
#> 32      4.982149    3.424685    1.5478032  0.006944302
#> 33      4.339169    3.452668    1.2513360  0.277952572
#> 34      4.930615    3.096831    1.3192506  0.070775659
#> 35      5.160625    3.486294    1.2012801  0.312545857
#> 36      4.013269    3.252978    1.5738379  0.724456187
#> 37      4.238015    2.999619    0.9771599  0.149221301
#> 38      4.001139    2.678994    0.7385114 -0.070310452
#> 39      4.185185    3.048466    0.7967965 -0.237276853
#> 40      4.687521    3.271688    1.2575825  0.172501837
#> 41      4.921600    3.489060    1.6025943  0.027873135
#> 42      5.060753    3.360482    1.4969859  0.059905906
#> 43      4.944743    3.054772    1.2901983  0.077428942
#> 44      5.071919    3.451943    1.9119629  0.133709144
#> 45      4.731359    3.042531    1.4473203  0.140226331
#> 46      5.833850    3.622025    1.9603710  0.087200587
#> 47      5.692124    3.750871    1.6799734  0.315151121
#> 48      4.929377    3.438733    1.5845159  0.464546701
#> 49      4.883575    3.246803    1.5386438  0.303732646
#> 50      4.419494    3.596714    1.7380197  0.565560599
#> 51      4.795261    2.314693    3.7990629  1.100771737
#> 52      5.401676    2.615534    3.7022700  1.101879141
#> 53      6.802093    3.152204    4.4331276  1.421171362
#> 54      6.083022    2.441584    4.1063721  1.096887145
#> 55      6.052667    2.585527    4.6178742  1.422844774
#> 56      6.426270    2.783705    4.6288781  1.398994007
#> 57      6.131356    2.520185    4.2384343  1.550051319
#> 58      6.407136    2.883483    4.4239041  1.492037427
#> 59      6.132242    3.236710    4.3694425  1.380064575
#> 60      5.781336    2.484103    4.2647023  1.252172037
#> 61      6.300680    2.812264    4.7729752  1.575162574
#> 62      6.735906    3.080122    4.4900285  1.624763570
#> 63      5.414032    2.686479    4.1911679  1.423840595
#> 64      5.684933    2.575063    4.2223137  1.189742444
#> 65      6.094437    3.504890    4.1289650  1.351369826
#> 66      5.037042    2.932755    3.7251191  1.238248972
#> 67      6.142954    2.862882    4.5755486  1.454480824
#> 68      6.709833    3.071116    4.6863533  0.979801807
#> 69      5.895783    2.875416    4.2337546  1.471993591
#> 70      5.858090    3.481163    4.0365299  0.998805124
#> 71      4.611001    2.392737    3.7095778  1.557175355
#> 72      5.475640    2.517146    4.0479404  1.218955954
#> 73      6.655027    3.021758    4.6240988  1.550800136
#> 74      5.696219    2.674717    4.0430234  1.411052834
#> 75      5.923893    2.924955    4.1598950  1.426670585
#> 76      5.374195    2.684463    3.8201579  1.412543980
#> 77      5.956318    2.843444    4.7607100  1.434516333
#> 78      5.889502    3.182364    4.5718201  1.447309731
#> 79      6.783533    2.236147    4.3678511  1.214534317
#> 80      6.755438    3.232567    4.9285580  1.519041723
#> 81      4.398775    2.211737    3.1764932  1.108725550
#> 82      6.724103    3.151147    4.8328635  1.427833552
#> 83      6.103665    3.452328    4.4683952  1.965817208
#> 84      6.714620    3.760334    4.9317555  1.719054294
#> 85      5.450538    2.624497    4.1059775  1.101748465
#> 86      6.219506    2.935380    4.4547051  1.366992810
#> 87      5.765926    2.166681    4.3739223  1.555901594
#> 88      5.899970    3.033595    4.2856361  1.742288505
#> 89      5.581818    2.643629    4.6434368  1.351250649
#> 90      5.487177    2.873935    4.1254540  1.447640209
#> 91      6.163431    2.853093    4.3373147  1.391868594
#> 92      5.692455    2.272752    4.2227790  1.369481637
#> 93      5.811874    2.956474    4.0153384  1.357559656
#> 94      6.429658    2.608554    4.2254654  1.270202774
#> 95      5.976732    2.944545    4.3376402  1.115147033
#> 96      5.356549    2.863296    3.6006800  1.077579691
#> 97      4.666279    2.086693    3.6279759  1.264085756
#> 98      5.621113    2.855417    4.6128592  1.435397688
#> 99      6.209803    3.088641    4.4814033  1.619719778
#> 100     5.510905    2.886208    4.3755503  1.320221754
#> 101     6.208322    2.944043    5.4267064  2.246916468
#> 102     7.087090    2.958868    5.6832321  1.976846304
#> 103     7.387799    3.455619    5.2681866  1.879263712
#> 104     6.229153    2.940436    5.3255341  2.147678591
#> 105     6.307080    2.968248    5.5148238  1.997696862
#> 106     7.452629    3.127955    6.2082858  2.176250659
#> 107     5.806001    3.173303    5.2343938  2.071049328
#> 108     6.521746    3.155000    5.5932100  2.091692373
#> 109     6.558175    2.714507    5.5429868  1.740767374
#> 110     7.036462    3.422186    5.7138688  2.031284866
#> 111     6.179885    3.260988    5.2658304  2.070092185
#> 112     6.451757    3.002252    5.2610444  1.838313031
#> 113     6.851649    2.957999    5.5791179  2.125109691
#> 114     8.430901    3.883596    6.8970678  2.431199020
#> 115     7.113339    3.507515    5.9947659  1.999050253
#> 116     5.946100    3.063429    5.2441755  1.558461995
#> 117     6.338337    3.040296    5.5878965  2.343254748
#> 118     6.962123    2.712172    5.8634983  2.153281055
#> 119     5.921617    2.627352    4.8661928  1.752229443
#> 120     6.332743    3.303128    5.6980797  2.241604662
#> 121     6.512692    2.441874    5.4256448  1.884418179
#> 122     5.760209    2.411699    5.2888761  2.099865418
#> 123     6.522570    2.540470    6.0784078  1.919214517
#> 124     5.978611    2.706133    5.3548118  2.221831257
#> 125     6.550382    3.190171    5.3495134  2.277179941
#> 126     7.555969    2.994589    6.2776856  2.110831379
#> 127     7.210033    3.296348    6.1178725  1.979642110
#> 128     7.076055    2.920659    5.4242193  1.514198803
#> 129     7.220820    3.214115    6.2683770  2.313540082
#> 130     7.077533    3.032830    5.9038882  2.009540365
#> 131     6.126979    2.900769    5.0534733  1.857838100
#> 132     6.336392    2.813514    5.5697073  2.018297035
#> 133     6.004611    2.182147    4.8279257  1.832572647
#> 134     7.417312    3.994178    6.0964850  2.038630963
#> 135     5.917091    2.623179    5.3231888  2.113978144
#> 136     6.224547    2.828757    5.5970253  1.971415582
#> 137     6.850740    3.102023    5.7797195  2.223604583
#> 138     7.008662    3.037148    6.2850112  2.284139479
#> 139     6.274473    2.770423    4.9553211  1.612147412
#> 140     6.842205    3.225005    5.7701422  2.021989988
#> 141     7.144392    3.318728    6.2763817  2.365074353
#> 142     6.533587    3.042034    5.3727515  2.335684518
#> 143     6.417883    3.395682    5.5655447  2.063131897
#> 144     7.315413    3.280929    6.1595532  2.419113810
#> 145     6.903853    3.246323    5.9104215  2.092791466
#> 146     6.495216    3.050020    5.5171673  1.963119445
#> 147     6.518774    3.436689    5.6119926  2.126249103
#> 148     6.280202    2.639479    5.0830610  2.009375437
#> 149     6.388017    2.945553    5.4188138  1.902261959
#> 150     6.462095    2.554274    5.8818168  2.444934948
#> attr(,"seed")
#>   [1]       10403         174   445293675  1931343245  1071229858  1126322308
#>   [7] -1241282712   417776412  -418361711   923475855 -1225310978    70680459
#>  [13]    58140182   -52491286 -1020948928   559162752  -310072563  1182641893
#>  [19]  1415003720  -353544896   798761919   874384219 -1939253161  1343997956
#>  [25]  1849691115  -567606761  1308298844 -2022453858   755217989   -56404265
#>  [31]  1290602219   991932981   403780784   490799361    39758591 -1025591959
#>  [37] -1344886051  -538178449  -395990347  -983582292  -136195132  -935131479
#>  [43] -1881951790 -1058959973 -1885545177  1393173100  1852576287  1756072154
#>  [49]   334232258   738530528 -1863855557  -779975170  2073231999  1511578914
#>  [55]   838438791   387958863  2060720174  -283419919    99494145  1485064538
#>  [61]  -275208702  1434885467   303634273  1252362161  1770857185  1347504018
#>  [67]  1005942471 -1292604986  1786829359 -1120611641  -257687293 -1409068143
#>  [73]  1849412880   531276323   967453240 -1726081163   653993432   714485552
#>  [79]   445353756   298752943  1193533988 -1394319086 -2112226916  1531306090
#>  [85] -1675442120  -677736819 -1910402550 -1699386026 -1335589432 -1172432112
#>  [91]  -953919552 -1828293659   338198950  -783119681 -1571279795 -1707720297
#>  [97]   472903416  1950244929   909154142  1317255698  -881497733 -1132262634
#> [103]  -321572206   538286042  2031128131   724877644   437393816   464042685
#> [109] -1359552221  1004500666 -1707024453  -752950153 -1397441828 -1295344424
#> [115]  -541771027 -1821768620   168328192   647439943 -1876571339  -220140103
#> [121] -1742150056  2061610002 -1677334598 -1440889265 -2123748759 -1718701065
#> [127]   516905177  -237722473 -1761903170  -970745285   -44840185   703093417
#> [133]   728794835   589189455  1350392442   321637406   521044222   718473710
#> [139] -1542578044   790959476 -1594104195   749665604  2099017048 -1109755916
#> [145]  1229537625  1681220738  -625383690 -1800740703  -814869268   605247917
#> [151]  -425882594  1443070761 -1479178764  1128584015  -990646153  -795011988
#> [157]  2028090425 -1027582401  -209837150 -1787119729 -1338245357   408220225
#> [163]  -435682735  -176287918  1436205392  1042689860   830755127  -391139303
#> [169] -1273149526 -1009474536  1748589920  1739478330   230050007    71956927
#> [175]  2058696277  1046606709   918505390 -1225205659   971859680  1239010374
#> [181] -1135723193 -1171810023  1710974950  1208845707   546456336   416330388
#> [187] -1496335001   409413161 -1177166286  1650743513 -2082639966  1179915976
#> [193]    93145423   169483655  -978003242  1020259119  -570427689   397876559
#> [199]   185367050 -1265906084  -753116188  -838940921 -1679939915  -103817305
#> [205] -1613822237  -576312677 -1777155169   656003767  -598545098   965599098
#> [211]  -489641184  -921437117  1195209081  1889076303   220049772  1452098201
#> [217] -1026771583   738203499  1659023995 -1019645894   255169933  1932041584
#> [223]  -330283926 -1509163270  1245103218  2008060153  1755712800   506678530
#> [229]  -169704595  1852321067  1428376063   465203313 -1206428684 -2132002885
#> [235] -1781935361  -100539048 -2014328109   635315235  1160832235   643352670
#> [241]   160434274   416103775  1644966180 -1245956688  1691322681   -64891581
#> [247] -1513887864    71365830  2025525113 -1979668284    98356040   743821835
#> [253]  1262926535 -1253827709   143341454 -1360354447 -1812899822   776607756
#> [259]   904895298  1384462628 -1473864417  -214283925  -851189245  1683980542
#> [265]  1887129516  -592128945   -64997931 -1325963942  1772051337  -958134529
#> [271]   -34818645 -1383607203 -1087059414 -1790791257  2138842640  1438842508
#> [277]   110199005 -1125480513 -1569504224 -1610587564  -672610039 -1710175962
#> [283]  -540407728    90478542   420491762  -712411697   179253537 -2007417514
#> [289]  1956002226  1811845913 -1413476745  -157449809  2087952737  -737850317
#> [295] -1873343648  1611478713  -421738281  1157727537  -242854291  1714112986
#> [301]  1679677321 -1164051695 -1723710936  1607344645  1630814187  1479807898
#> [307]   408425985   680580289  2100111366 -1279472844   741945750 -1116556096
#> [313]    66258119  1209732446   637501438  -753637836 -1849380237  1174324094
#> [319]  2131082752 -1192906100 -1453097705  -424486658 -1979555792  1395302455
#> [325] -1432984279  1118269938 -1007717316  -846390571  -535218683  -514679982
#> [331]   445341721  1667516039 -1955149723   660559530  -267958640    14686717
#> [337]   459249459  -368222913  -728458710  1667389738   -56956763  -806693702
#> [343]  1107053782 -1394162220  -903829506   150242040 -1401476034  -988666883
#> [349]  -197522304  -929083219 -1931506989  -690637150 -2044232038  1891014029
#> [355]  -166774174  -538162126  -408534187  2018848420  2057177435  -856897793
#> [361]   734868065  -259870202 -2062787978   -23853012 -1143409522   356894846
#> [367] -1722563616  1270911960  1403934106 -1667067333  -702888906   128315272
#> [373]   346917364  1464396030  2074129819 -1025392312   372169696 -1226684213
#> [379]   303009639  -394118505  1466443026  1947046426 -1110800786 -1841583757
#> [385]  -191388985  1749726344 -1173764246   891863222  -895853399   558285101
#> [391]  -554947445  1444433448  1715018450  1580347834 -2126139007 -1275121275
#> [397]  1915853400 -1016517817 -1243270946  1963334029  -230431153  -746956335
#> [403]  -333335259 -1128452666  1984265974  1262323975   321375164  1284876746
#> [409]   215273560  -160568363  1710489717  -646259141 -1164835269  -410007314
#> [415]  1865233391  -867780479  1584976596  1500011182    81534472   197456092
#> [421] -1037098412   407832597  1915580887 -1217899794  -622045141  -458248544
#> [427] -1908801926  1108388526   699347852   930745104 -1410110722  1403054181
#> [433]  1292371022   736374145  2136959165 -1805288194     3626879   441102404
#> [439]   775864636 -1122673733   285494040 -1138630357  1915140934  -454041165
#> [445]  1962189528 -2058293710  1932352349  1885956580  1487282385  -613280966
#> [451]  1164428168  -742927554  -357501494 -1326175915 -1972302474  2106268833
#> [457]   190334495   234675238  1774615147   -70548202  1600772261  1680012369
#> [463]  1156944740  1346862959   814221001  2129242312   185608288   298428090
#> [469]   511127393 -1973256661  1670945100 -2106989228  -920149575  1819100986
#> [475]  1915654811   121410612  1667073584  -510828339 -2089148607   569555916
#> [481]  1671530600 -1057991597  1964223413   523833514 -1058825063  -856762946
#> [487]  1484650268 -1367508969  1890082665  -473645728   605164637 -1726100722
#> [493] -1375189906  1912716756   706462219 -1396860238   976860729   932687124
#> [499]   991368215  -658015942 -1067689964   158802598 -1614420065   913208701
#> [505]   164175724  1508479748 -1793821200   491964811   774025301  1545918822
#> [511]  -946729628  -703316042   201625192  2072716555 -1672156329   431683371
#> [517]  1328603196  2055292180  1591837333   864391486  1265764238  1442467871
#> [523]   255693404  1067092789  -686952932 -1299638382  2045566007  -130677948
#> [529] -1520413159  -203420739 -2091629559  1057135946  -702265225  1711986347
#> [535] -2050086895   723871673  -790212943  -827747880  -327604815  1052022314
#> [541]   977044688 -1979179257   669110963 -1453250506  2078625413  -787429807
#> [547]   496350816  1580502116   684709750   -59753965   829054984   385653026
#> [553] -1080997752  -444936963  1922442846  1316342392   925794305   721396856
#> [559]  1998644565 -1052959065  -514715615  -110653746  1001443258  -764143271
#> [565]  -618216886 -1522451602 -1584958931  1498876023   793419273   521286152
#> [571] -1615843935    73404945  1813434239   -83982801  -106408794  2027408972
#> [577] -1645376293 -1577562079   876106613   989210780  1508495780  1318916483
#> [583]  -408131585   179287739 -1718812376    30211166  -752760804 -1941158393
#> [589]  -785590115  1119872630  1981303367   -97915192  -874363705  -197949344
#> [595]  -665467222 -1687851530   863357964   423589014   237162524  1077422753
#> [601]   549950807 -1633882566  1183346086  1780989289  -677329104   940601185
#> [607]   706553167  -151207405   821383085 -1325337252   303439233 -1230037192
#> [613]  -716601064 -2095016736   493739987  1958896868 -1140650008  -553675236
#> [619]   582128919  1611363834 -1422118398   725977646   737100074   -46177562
#> [625] -1690794529   564711988