Simulate Responses from a Multivariate Linear Model

Simulate one or more sets of responses from a Multivariate Linear Model (mlm) object.

Usage

# S3 method for mlm
simulate(object, nsim = 1, seed = NULL, ...)

Arguments

object

a mlm object, typically the result of calling lm where the response is a matrix.

nsim

number of replicate datasets to simulate. If nsim is greater than 1, the output is arranged in a 3D array.

seed

an object specifying if and how the random number generator should be initialized (‘seeded’). Either NULL or an integer that will be used in a call to set.seed before simulating the response vectors. If set, the value is saved as the "seed" attribute of the returned value. The default, NULL will not change the random generator state, and return .Random.seed as the "seed" attribute, see ‘Value’

...

additional optional arguments.

Value

A matrix of simulated values for the response (or an array, for nsim greater than 1)

Details

A simulate function for mlm objects, which simulates one or more sets of responses from a Multivariate Linear Model (mlm) object. If multiple sets of responses are requested, they are returned in a 3D array, with simulation number along the third dimension.

The weights argument is currently ignored -- a constant variance-covariance matrix assumed for mlm.

See also

Author

David Warton <david.warton@unsw.edu.au>

Examples

# fit a mlm to iris data:
data(iris)
iris.mlm=lm(cbind(Sepal.Length,Sepal.Width,Petal.Length,Petal.Width)~Species,data=iris)
# simulate new responses:
simulate(iris.mlm)
#>     Sepal.Length Sepal.Width Petal.Length  Petal.Width
#> 1       5.874034    3.794902    2.1836820  0.648809803
#> 2       4.844432    2.665505    1.6064944  0.117537652
#> 3       4.704402    3.785614    1.5163707  0.125529281
#> 4       5.789551    3.149518    1.6429360  0.186730420
#> 5       4.950210    3.265321    1.5186447  0.264313693
#> 6       5.113463    3.443978    1.6515124  0.514278561
#> 7       5.540633    3.962108    2.1090488  0.449297021
#> 8       5.493554    3.038241    1.6938034  0.035723618
#> 9       4.500842    3.526181    1.3463866  0.356111721
#> 10      5.043088    3.250134    1.7044769  0.174966296
#> 11      4.997800    3.554038    1.2991926  0.343165309
#> 12      5.614934    3.686363    1.9905298  0.575969078
#> 13      5.188701    3.333677    1.3507264  0.232437745
#> 14      5.753152    4.023175    1.5128243  0.443790317
#> 15      4.534817    3.270247    1.1521567 -0.052475522
#> 16      5.969652    3.920572    2.1234386  0.701332726
#> 17      4.057270    2.768477    1.0311796  0.005630903
#> 18      5.123670    3.773714    1.3857623  0.290753789
#> 19      4.652336    3.626815    0.9974827  0.059938272
#> 20      5.587669    3.315268    1.5233360  0.185295415
#> 21      4.413398    3.476650    1.3823783  0.555149406
#> 22      4.816361    3.572207    1.1221683 -0.189048346
#> 23      5.339580    3.109738    1.8033028  0.182055464
#> 24      5.853404    4.111873    2.0251529  0.568193341
#> 25      5.588893    3.295652    2.1325931  0.300645344
#> 26      4.040344    2.968678    0.7400400 -0.006384120
#> 27      5.562399    3.596604    2.2161952  0.485495364
#> 28      4.533453    3.728611    1.3219442  0.397050836
#> 29      6.118582    4.055947    2.5439821  0.360464079
#> 30      5.525496    3.576468    1.7272153  0.012326998
#> 31      4.770202    3.004246    0.9177282  0.094517642
#> 32      5.184344    3.335099    1.6648016  0.335582182
#> 33      5.131619    3.287776    1.3057955  0.141298308
#> 34      5.026381    3.425277    1.1303592  0.205884427
#> 35      5.013616    3.389301    1.8994638  0.432835984
#> 36      5.246444    3.596985    1.7138815  0.414631106
#> 37      4.374913    2.727091    1.2614367 -0.025423238
#> 38      5.509722    3.641875    1.4182597  0.372988376
#> 39      4.593504    3.347447    1.3581127  0.147855002
#> 40      4.906898    3.788488    1.4411858  0.283137369
#> 41      4.176635    3.171602    0.7065201  0.001448908
#> 42      5.628146    3.635211    1.5871807  0.498199423
#> 43      4.580953    3.124337    1.1579366  0.222269169
#> 44      5.168327    3.416716    1.6602774  0.536839815
#> 45      5.009732    3.672188    1.5646542  0.167257290
#> 46      5.682668    3.719913    2.1912232  0.730318140
#> 47      5.019681    3.243336    1.6864580  0.303185107
#> 48      5.337929    3.633601    1.9233950  0.356944628
#> 49      4.385671    3.521298    1.2748834  0.255942640
#> 50      5.744306    3.721348    1.6928181  0.327227414
#> 51      5.633736    2.994567    4.0197714  1.570015308
#> 52      6.070740    2.793912    4.1380369  1.300075306
#> 53      5.034823    2.275263    3.4289591  1.457074751
#> 54      6.872454    2.972533    4.5499160  1.297115839
#> 55      5.836235    2.424622    3.8089993  1.240294968
#> 56      5.423481    2.483145    3.7094614  1.224271223
#> 57      5.715391    2.360901    4.0992238  1.360022023
#> 58      5.533271    2.673380    3.8946112  1.511060581
#> 59      6.378660    2.947853    4.4985057  1.419558344
#> 60      5.090712    2.353913    3.5309870  1.373107263
#> 61      7.253925    3.414502    5.0985103  1.763990773
#> 62      6.075283    2.863009    4.4821142  1.602600006
#> 63      6.328380    2.818488    4.7312858  1.348081454
#> 64      6.093600    2.554147    3.8614728  0.719504388
#> 65      4.630260    2.392110    3.1325697  1.039771912
#> 66      5.044882    2.538889    3.7232373  1.161288596
#> 67      5.668453    2.441636    4.1958916  1.409297065
#> 68      5.253820    2.373557    3.7808064  1.312018280
#> 69      5.278299    2.749286    4.0638863  1.442299123
#> 70      5.943622    2.391557    4.3822934  1.296159787
#> 71      5.522594    2.709941    3.6959762  1.428396333
#> 72      5.875354    2.453050    4.6031905  1.495042327
#> 73      5.099504    2.500940    4.0427648  1.066706147
#> 74      5.975881    2.575535    4.1552584  1.102773882
#> 75      5.476507    2.492902    4.1791064  1.379503181
#> 76      6.543831    2.616198    4.9855438  1.571418218
#> 77      5.755657    3.357979    4.6333958  1.745504324
#> 78      5.672168    2.455171    4.7028136  1.117238261
#> 79      6.282994    2.490720    4.2128881  1.467097408
#> 80      6.132645    2.552603    4.7562973  1.371606631
#> 81      7.015896    3.201523    5.1179360  1.598578351
#> 82      6.280175    2.848900    4.3958252  1.366742584
#> 83      5.298108    2.581637    4.1662094  1.695336144
#> 84      6.291270    3.259799    4.3434289  1.258077222
#> 85      5.650732    2.831897    4.4574016  1.363831493
#> 86      5.157563    2.643309    3.7423707  1.334456121
#> 87      6.641818    2.857988    4.3853299  1.229401618
#> 88      5.099571    2.355795    4.1661264  1.133287410
#> 89      5.027058    2.387532    3.6938263  1.086445858
#> 90      5.813833    2.745899    3.8624814  1.211387051
#> 91      5.794973    3.058997    3.9126712  1.361916392
#> 92      6.038407    2.791675    4.4848228  1.610996359
#> 93      6.126830    2.753508    4.0253926  1.062970395
#> 94      5.537825    2.368936    4.4311616  1.189613985
#> 95      5.643935    2.773207    4.1262196  1.315030726
#> 96      6.645428    3.245020    4.5487904  1.500933675
#> 97      5.899132    2.222750    4.3991983  1.181618638
#> 98      6.507821    3.355169    4.8297509  1.682941239
#> 99      6.044098    3.304192    4.0966974  1.637386274
#> 100     5.764607    2.898704    4.1458880  1.559018538
#> 101     5.996176    2.995484    5.0817801  1.860154654
#> 102     6.581437    2.604138    5.2200356  2.149833191
#> 103     6.032696    3.157653    4.8949287  2.096748431
#> 104     7.625262    3.752928    6.4522476  2.096691340
#> 105     6.516927    3.187310    5.5157168  1.972737950
#> 106     5.693066    2.809612    5.3548603  2.147512290
#> 107     6.582523    3.040477    5.6320265  1.837942946
#> 108     6.298620    2.952042    5.5095540  1.863006338
#> 109     7.585954    3.454844    6.2991356  1.930185101
#> 110     6.269012    3.255855    5.4017932  2.037678380
#> 111     7.498705    3.295279    5.6349859  1.979417394
#> 112     6.633990    2.532407    5.3860678  1.817933492
#> 113     7.016588    2.959852    6.0184732  2.285299992
#> 114     6.318198    2.712409    5.3017586  2.003290583
#> 115     6.390914    3.291206    5.0369682  2.038758312
#> 116     6.180984    2.594022    5.3986453  1.992393079
#> 117     6.177153    3.289008    5.0548582  1.853457794
#> 118     6.802877    2.729344    5.9915165  2.030533624
#> 119     6.902266    3.104932    6.1279931  2.024621457
#> 120     6.580589    2.455644    5.8012166  1.922921866
#> 121     6.974910    3.222913    5.8082757  2.350220125
#> 122     7.230371    3.267268    6.0897552  2.020196851
#> 123     7.098491    3.092535    5.6403793  2.143326543
#> 124     6.670553    2.587051    6.1610026  2.262456287
#> 125     6.284048    2.851804    5.1464267  1.846072432
#> 126     7.303408    3.419406    6.2858532  2.230990863
#> 127     7.364382    3.124378    6.3382372  2.250779586
#> 128     7.058404    2.666769    5.9135833  2.057890734
#> 129     6.125704    2.625697    5.0637804  1.870844989
#> 130     7.177941    3.018248    6.1431695  2.104594612
#> 131     6.634893    3.227100    5.6405938  2.052616063
#> 132     7.787935    3.355289    6.4939070  2.238911420
#> 133     7.115195    3.060831    5.3262507  1.860188199
#> 134     7.192462    3.246958    6.3071241  2.231322875
#> 135     6.602475    2.936885    5.2534588  1.873040421
#> 136     6.770606    2.858313    5.7758576  2.125671226
#> 137     6.825718    3.750064    5.8458594  2.221010334
#> 138     7.000965    3.373179    5.7378679  2.254501114
#> 139     6.378829    2.985379    5.3961129  2.049106390
#> 140     6.838848    3.493481    5.5770412  1.954229838
#> 141     7.033694    3.472891    5.6695080  2.125002264
#> 142     6.646510    2.999485    6.1463255  1.761448389
#> 143     6.557618    3.018369    5.2545426  2.184725385
#> 144     5.837368    3.003464    4.6891070  1.843538409
#> 145     6.382185    2.764106    5.1911538  1.947116529
#> 146     6.796228    2.941732    5.4237831  1.850351249
#> 147     7.274669    3.244338    6.4927931  2.319180263
#> 148     6.185488    3.217795    4.9315020  2.224483871
#> 149     6.205056    3.189287    5.1832425  2.175169274
#> 150     6.563431    3.272319    5.7572582  2.012484335
#> attr(,"seed")
#>   [1]       10403         527 -2000719847  -472489860 -1707854064   601706559
#>   [7]  -751836274 -2140291320 -1527625769  1943825166   942407783  1071134026
#>  [13]  1322536133  1251434364   228373662   942345967 -1066229774  1613797339
#>  [19]   355116066  -804753240   448587841 -2086400738  -582981298  1175523817
#>  [25]   179478229  2030277151  1155219411  -315357830  -264969529 -1469923256
#>  [31]  -636323630  1077993486    35753438  1494155093  2134033360 -1150385906
#>  [37]   941910431  2029515710  1939534366 -1241103031   -62696812  -139337393
#>  [43] -1911296156 -1761335852  1609915237 -1889234232 -1791763992   651562414
#>  [49]  1150468100  -811742001   567510795 -1039562131  1974686642 -1647064967
#>  [55]  1758220880  -430422766 -1923257094   992215925  -487422228 -1861493030
#>  [61]  1863668533  1964096487  1151248544  -218488731  1304368960  1421919691
#>  [67]  -330045504 -1610649288 -2010459305  -892399524   687688959   413544661
#>  [73] -1979366613  -956963541  -644571454  1197964785  2014678164   165707119
#>  [79] -1086284859  1467701248  -956765236  1035956891   810780633   173993094
#>  [85]  1487028958 -2111362569  -604029864  -800183980  1433381698  -284080540
#>  [91]  -606886829  1763596260  -778176215 -1690510347 -1145127859   -43511807
#>  [97]  1357532302 -1451550625  -940597710 -1356671754  -617264861 -1164205309
#> [103]  -281671914 -1047105055  2145974137  1766067603  1476723468 -1006309189
#> [109]   588253035  -742271307 -1906693056 -1275866566  1141571497  -181020816
#> [115] -2087280139 -1789149403   -76512094 -1317111859  2051330076  1052446993
#> [121]    61606697   787051141  -742764406  1105842591 -1294569244  -400907595
#> [127]   593128168 -1811390897  1291723893   824560808 -1814353139  -610684493
#> [133]   754918654  1484638361  -297519714   400162362   166529507 -2127723025
#> [139]  -999710540  1462260948  1860819194  -116827672   395676327  -655119592
#> [145] -1370067673   426628533 -1818745906 -1968985201  1395240224 -1630215005
#> [151]  1862267538 -1846542456 -1713083130  -911192800   840688022  -776782451
#> [157]  1945218461  1184451460 -1326934330 -1169537908 -1618462479 -2060103717
#> [163]   160655555  -382866620  1892523350  1628243266   557241823  -195926660
#> [169]   680540695  1572162538    27291520  1306855015 -1537916039   681092098
#> [175]   382407114 -1641024594  1750254141 -1577161431  1464285940   184445860
#> [181]  -151036213  1399150730 -1876189495  1018858541 -1180912604 -1889462839
#> [187] -1399136889   479325959 -1124518149 -2097734894  1775821972  1995875028
#> [193]  1130367642  1353991353  1437801808  2062609090 -1466992730  -760919692
#> [199]  -415085059   445657451    23401087 -1470454006 -2086773860  -477283542
#> [205] -1874733228 -1748228218 -1268116523   613671612   540258420 -1008513414
#> [211]  -716864563  2052815103   -30388425  -757024717  1152921310  1859768967
#> [217]  -371133631 -1781698369  -416839590 -1656074145  1054439016 -1131151866
#> [223] -2000757325  1074442549 -1677459323   362526040  -248371556   479609079
#> [229]   338877913  1445107575  2048341353   586629476  1717626508 -1987573130
#> [235] -1154267493  -860479638 -1225061834   433203899  1057918364   -60236413
#> [241]  -337769705  1253993388   949969542  2025751399 -2026639653 -1909725423
#> [247]  -673061972  -972445051   106949152 -1880772446  1489463043  -737140079
#> [253]   933659444   790395197  -104024494  -719149212 -1928172695  1893047885
#> [259]  1026626682 -1654455651   199998372 -2137982488   643921609  1678407920
#> [265]   913689108  1578729009    64159249   -52233548  -522909207  1593298371
#> [271]  -141830114 -1308612030 -2146981746 -1378984415   365166390  -510205416
#> [277]  1669324354  -519485234   457850466   289338040  1557344107  -907060981
#> [283]  1814798758   750494469  -352113281 -2138008076   134377281 -1758184408
#> [289]     3994975  -608065795  -104183616   570098745 -1430680210 -1162222043
#> [295]  -175791097 -1577922390 -1637063709 -2030170090 -1918499190  2072147928
#> [301]   748828944 -1568938721  -840371401  1494440353  -381099944   726325452
#> [307] -1178968945   797980717  -679719237   285148697   120991042 -1647188584
#> [313] -1950272218   -51843544  1465502115 -1047484594  1215679046   868131667
#> [319] -1037190058    28821932  -936625781   196394652   408730036  1389616004
#> [325]   875097028  -486053950  1352342674 -1874865051  -742839769   -52176711
#> [331]  -446537615  -565372622  1169309955   249068359 -1203363647   918191416
#> [337]  -723869348   499903842  1743404340 -1189573056 -1038770046 -2067704522
#> [343]  -424171307   976393956 -1450400277   375168576  -663319907  -770367110
#> [349]  1662096509   364293173   -38098227  1977507931  -331592184 -1347757285
#> [355]   -97912872  -248171933  -666815495  -170048912  -322735567  1726232354
#> [361]   355991354  -361278626 -1915809150  -122484715  -970782924  -534863244
#> [367]  2134010814   221680716   506320952  2117922683  1515640660 -1455442972
#> [373]   982110460  1081557155 -1804122264      622917 -1299493443  1152169372
#> [379]  1803688851  1114861481  1122637341   754253367 -1595222342  -105344212
#> [385]  1413761668  -469937371  -157285324  1884911122 -1513966935  1834209682
#> [391] -2105213842  1206255499  1804030775 -1000011526  -591844678    12473716
#> [397]   -79646500  1499348918   614726754   961162232   212286325 -1315485601
#> [403]  1112377375  -810622099   848525434  1813358512 -1057865437  -141827251
#> [409] -1428584950   786794321 -1230285852 -1764565832 -1973359172    -9339152
#> [415] -2069382724   101268620  1630748492  -795614395 -1385290458  1056668752
#> [421]  1038220104  -834611144   319877706   276329894 -1256208402  2074525233
#> [427]   894146798  1705641597  1639684701   166076328 -1566511905  1862224943
#> [433]  -890609110 -1656528714  1413074684  1105151132  -329496921  1569289453
#> [439]  1773490437  -771665268 -1997996337  -875119004 -1524471142    -4651685
#> [445]  -634328938  -616730009  -624778649  1250505317  1927881132 -2090638921
#> [451]   908391605  -608812698  -645722482  -326421626   307231074  -295887168
#> [457] -1824502856  -142065146  -456083017 -1981697590 -1517956254   977866261
#> [463]  -705473629 -1691128154 -1744595825  1274580116  -650796592 -1149584902
#> [469] -1696255598   780005478 -1246279562 -1042779574   476801736  -388338010
#> [475]   890921499   700863734   199708243 -1982707650  -898365214    45508092
#> [481]  1053884582  1847209297 -1565692816  1970754007   942011486  1019388267
#> [487]  1811979977   586047791 -1962900325  -771367915 -1377805454  -742921261
#> [493]   963474116 -1884813401  -174225147  -782703882   641056808 -1158656071
#> [499]   898520787   492525397  1838491836  1246601885  1756661511 -1491550694
#> [505] -1967630682  1428196724 -1794187810   856641041 -1105449568 -1802294098
#> [511] -1845273302   136777559   764119497  -659034417  -957576797  -988454931
#> [517] -1283988375  2051369484  1002531299   477985665 -1378393493 -1844042322
#> [523]   587394882   -90569420 -1932289168   918497936  1251898853  2136451256
#> [529]  1910634380  1424873901  -377519856 -1224807495  1672115368  -906875573
#> [535] -1649391159  -724128970   900963913  2034374745  -311576758    54861139
#> [541] -1719209856  -223578058 -1745882862  -618591140   142281997  -782702350
#> [547]  -869875836  -711597168  1704262531   568240911  1015796347  -447064650
#> [553]   247123976    49410216  1796931416  -100669545 -1181629321  1913016438
#> [559] -1923065518 -2039012696  1538068775  2119805166  1995979983 -1841299540
#> [565]   286000445  1019792995  1324908129  -880888566 -1370288780 -1259030752
#> [571]  1625769896  -679822993  1139684232  1908704088 -1266538853 -1804062321
#> [577] -1711228341  -540012566   670898655  -231198053  1153308106  -322330306
#> [583]   133592896 -1398413666  -276377450  -981207449  1052283752  -537435962
#> [589]  1708088725 -1232043944    27237550 -1623726756  1622655698 -1956572534
#> [595]  2060312217   -19067819 -1407602690   196958199   899868809  -344787600
#> [601]   542774944  -200958018  -602603520 -1886949907 -1451265679   609561398
#> [607]  -379750194  -295008433 -2092947795 -1022107068   631898662 -2000478752
#> [613]  -574686189   824819099 -1248577609  -314942807 -1652645316  -490313601
#> [619]  -154624435 -1198697308  2067163072 -1600873188 -1894516561 -1916007351
#> [625] -1314440535 -1175128911